% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{bonjour}
\usage{
bonjour(nom = "Maria", verbose = T)
}
\arguments{
\item{nom}{char. Name of the person to say boonjour.}

\item{verbose}{bool. Print in console}
}
\value{
string
}
\description{
Description.
}
\examples{
bonjour()

}
