% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour2.R
\name{bonjour2}
\alias{bonjour2}
\title{Say "Bonjour 2 times !" with a name.}
\usage{
bonjour2(nom = "Maria", verbose = TRUE)
}
\arguments{
\item{nom}{Name of the person to say " Bonjour !" (character).}

\item{verbose}{Print in console (boolean).}
}
\value{
A character string "Bonjour! Name Bonjour! Name"
}
\description{
Say "Bonjour 2 times !" with a name.
}
