% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createcanopy.R
\name{createcanopy}
\alias{createcanopy}
\title{Create the canopy}
\usage{
createcanopy(inventory)
}
\arguments{
\item{inventory}{Input inventory containing the crown diameter (CrownDiameter,
in m) (calculable with the addtreedim function) of each tree
(data.frame)(see the inputs formats and metadata in the
\code{\link{LoggingLab.html}})}
}
\value{
A dataframe with a column 'Crowns' containing the circles
(sfc_POLYGON) as trees crown, with their diameter filled in, representing
trees from the sky.
}
\description{
Create a vertical projection of the the crown of each tree
}
\examples{
data(Paracou6_2016)
Paracou6_2016 <- dplyr::slice(Paracou6_2016, 1:10) # inventory reduction

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- createcanopy(inventory)

# The small ones first so that they are plotted below the big ones
inventory <- dplyr::arrange(inventory, TreeHeight)

library(ggplot2)
ggplot() +
  geom_sf(data = getgeometry(inventory, Crowns),
          aes(alpha = TreeHeight),
          fill = "forestgreen") +
          labs(alpha = "Tree height")

}
