% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreekDistances.R
\docType{data}
\name{CreekDistances}
\alias{CreekDistances}
\title{Relative vertical and horizontal distances from nearest channel network
(Paracou plot 6)}
\format{
A list of 2 large RasterLayers \strong{with a crs  in UTM}

To generate vertical and horizontal distances from nearest channel network,
of any plot:
\code{\link{CreekDistances}} in 'articles' folder of the package
}
\usage{
CreekDistances
}
\description{
Relative (vertical (\emph{distvert}) and horizontal (\emph{disthorz})) distances (in m)
from nearest channel network of plot 6 of Paracou (French Guiana) (LiDAR, 1 m
resolution, 2016)
}
\keyword{datasets}
