% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacheckformat.R
\name{datacheckformat}
\alias{datacheckformat}
\title{Check & format your inventory data for the package "Maria"}
\usage{
datacheckformat(data = "Paracou6_2016", verbose = TRUE)
}
\arguments{
\item{data}{(dataframe) Your inventory data or the "Paracou6_2016" test data}

\item{verbose}{(boolean) Print in console}
}
\value{

}
\description{
Check & format your inventory data for the package "Maria"
}
\examples{
datacheckformat(data=Paracou6_2016)
}
