% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DemParacou.R
\docType{data}
\name{DemParacou}
\alias{DemParacou}
\title{DemParacou}
\format{
A RasterLayer
}
\usage{
DemParacou
}
\description{
Digital terrain model (DTM) of the plot 6 of Paracou (1m resolution, LiDAR campaign of 2016) (French Guiana)
}
\keyword{datasets}
