% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMParacou.R
\docType{data}
\name{DTMParacou}
\alias{DTMParacou}
\title{Paracou Digital terrain model (DTM)}
\format{
A RasterLayer
}
\usage{
DTMParacou
}
\description{
Digital terrain model (DTM) of the plot 6 of Paracou (1m resolution, LiDAR campaign of 2016) (French Guiana)
}
\keyword{datasets}
