% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAccesExplArea.R
\name{FilterAccesExplArea}
\alias{FilterAccesExplArea}
\title{FilterAccesExplArea}
\usage{
FilterAccesExplArea(
  exploitpolygones = ExploitPolygones,
  maintrails = MainTrails,
  winching = winching,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{exploitpolygones}{Accessible area of the inventoried plot
(default: \code{\link{HarvestableAreaDefinition}}) (sf polygons data.frame)}

\item{maintrails}{Main trails defined at the entire harvestable area (sf polylines)}

\item{winching}{no cable or grapple = "0", only cable = "1", grapple + cable =
"2"}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
Accessible area  of the inventoried plot according to selected logging scenario (sf polygons data.frame)
}
\description{
FilterAccesExplArea
}
\examples{

 AccessPolygones <-  FilterAccesExplArea (exploitpolygones = ExploitPolygones,
 plot = Plots,
 maintrails = MainTrails,
 scenario = scenario,
 fuel = fuel,
 diversification = diversification,
 advancedloggingparameters = loggingparameters())

}
