% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurereserve.R
\name{futurereserve}
\alias{futurereserve}
\title{futurereserve}
\usage{
futurereserve(
  inventory,
  speciescriteria = SpeciesCriteria,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{your inventory (see the inputs formats and metadata in the \code{\link{vignette}}) (data.frame)}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in the
same format of \code{\link{SpeciesCriteria}} (data.frame)}

\item{otherloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
your inventory with selected future and reserve trees, for your logging criteria (\code{\link{SpeciesCriteria}})
}
\description{
futurereserve
}
\examples{

data(Paracou6_2016)
data(DemParacou)

inventory <- ONFGuyafortaxojoin(addtreedim(cleaninventory(
inventorycheckformat(Paracou6_2016))))

harvestableOutputs <- harvestable(inventory, diversification = TRUE,
 specieslax = FALSE,
DEM = DemParacou, plotslope = PlotSlope,
otherloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, type = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters(), VO = 30,
 HVinit = HVinit)$inventory

futurereserve(inventory)

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}}, \code{\link{loggingparameters}}
}
