% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genaccesspts.R
\name{genaccesspts}
\alias{genaccesspts}
\title{Generate access points for machine}
\usage{
genaccesspts(
  topography,
  machinepolygons,
  maintrails,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer \strong{with a crs in UTM})}

\item{machinepolygons}{Accessible zones for machines of the inventoried plot
(default: \code{\link{harvestableareadefinition}}) (sf polygons data.frame)}

\item{maintrails}{Main trails defined at the entire harvestable area (sf
linestring \strong{with a crs in UTM})}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
list of 2 elements :
\itemize{
\item \emph{PartMainTrails} available area to draw access points of maintrail for each PU (prospection
unit) (sf or sfc)
\item \emph{AccessPointAll} a draw of access points of maintrail for each PU (prospection
unit) (sf or sfc)
}
}
\description{
Accessory function to generate access point according to operating-machine area.
}
\examples{

\dontrun{
set.seed(1)
data(MainTrails)
data(HarvestableAreaOutputsCable)
data(DTMParacou)

accesspts <- genaccesspts(topography = DTMParacou,
machinepolygons = HarvestableAreaOutputsCable$MachinePolygons,
maintrails = MainTrails,
advancedloggingparameters = loggingparameters())
}

}
