% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmask.R
\name{getmask}
\alias{getmask}
\title{Get mask}
\usage{
getmask(inventory, topography)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
vignette) (data.frame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR,
1m resolution)
(RasterLayer \strong{with a crs in UTM}) (See \code{\link{DTMParacou}})
We advise you to generate your raster with Qgis rather than with the
'raster' package on R.}
}
\value{
The inventoried plot mask (SpatialPolygonsDataFrame
\strong{with a crs in UTM})
}
\description{
Build a mask with the intersection of the topography and the spatial inventory.
}
\examples{
data(Paracou6_2016)
data(DTMParacou)
mask <- getmask(Paracou6_2016, DTMParacou)

}
