% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaDefinition.R
\name{HarvestableAreaDefinition}
\alias{HarvestableAreaDefinition}
\title{HarvestableAreaDefinition}
\usage{
HarvestableAreaDefinition(
  topography,
  verticalcreekheight,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (Large RasterLayer)}

\item{verticalcreekheight}{Relative elevation from nearest channel network
(Large RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with:
\itemize{
\item A collection of polygons defined as 1 : harvestable area, 0 :
non-harvestable area (sf polygon)
\item A raster with slope (in radians) characteristic of the studied plot
(Large RasterLayer)
}
}
\description{
HarvestableAreaDefinition
}
\examples{
data(Plots)
data(DTMParacou)
data(VerticalCreekHeight)


HarvestableArea <- HarvestableAreaDefinition(topography = DTMParacou,
                                             verticalcreekheight = VerticalCreekHeight,
                                             advancedloggingparameters = loggingparameters())

plot(HarvestableArea[[1]])

}
