% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaOutputsCable.R
\docType{data}
\name{HarvestableAreaOutputsCable}
\alias{HarvestableAreaOutputsCable}
\title{'harvestableareadefinition' outputs when the cable will be used
(Paracou plot 6)}
\format{
list
}
\usage{
HarvestableAreaOutputsCable
}
\description{
Harvestable and non-harvestable zones of Paracou plot 6 (French Guiana)
}
\details{
A list with:
\itemize{
\item 'HarvestablePolygons': a collection of polygons (sfc_MULTIPOLYGON) defined as:
1 : harvestable area,
0 : non-harvestable area
\item 'PlotSlope': a raster with slope (in radians) characteristic of the studied
plot (Large RasterLayer)
\item HarvestableArea : the harvestable area in hectares (double)
\item MachinePolygons : a collection of polygons (sf (sfc_POLYGON)) defined as:
1 : accessible machine area,
0 : non-accessible machine area
}
}
\keyword{datasets}
