% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestablePolygons.R
\docType{data}
\name{HarvestablePolygons}
\alias{HarvestablePolygons}
\title{Harvestable zones (polygons) of Paracou plot 6}
\format{
sf (sfc_POLYGON)
}
\usage{
HarvestablePolygons
}
\description{
Harvestable and non-harvestable zones of Paracou plot 6 (French Guiana)
}
\details{
A collection of polygons defined as:
1: harvestable zone
0: non-harvestable zone
}
\keyword{datasets}
