% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorycheckformat.R
\name{inventorycheckformat}
\alias{inventorycheckformat}
\title{Check your input inventory data for the  package "Maria"}
\usage{
inventorycheckformat(inventory)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}
}
\value{
Stop the function if the format is not the one required.
}
\description{
Check your input inventory data for the  package "Maria"
}
\examples{

data(Paracou6_2016)
inventorycheckformat(inventory = Paracou6_2016)

}
