% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingLab.R
\docType{package}
\name{LoggingLab}
\alias{LoggingLab}
\title{LoggingLab-package}
\description{
Logging simulator
}
\details{
This package allows to simulate the different low impact logging versions in French Guiana.

The LoggingLab package provides n functions:
treeselection,  and .
}
\section{LoggingLab functions}{

loggingsimulation

loggingparameters

scenariosparameters

inventorycheckformat

addtreedim

treeselection

commercialcriteriajoin

harvestable

selected

futurereserve

treefelling

createcanopy

treefromthesk

directionalfellingsuccessdef

felling1tree

rotatepolygon

getgeometry

timberharvestedvolume

exploitablefuelwoodvolume
}

