% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingSimulationOutputs.R
\docType{data}
\name{LoggingSimulationOutputs}
\alias{LoggingSimulationOutputs}
\title{Logging Simulation Outputs}
\format{
list
}
\usage{
LoggingSimulationOutputs
}
\description{
Outputs of the logging simulation on the plot 6 of Paracou (French Guiana)
}
\details{
With the parameters:
scenario : manual
objective : 20 m3/ha
fuel : 2
diversification : TRUE
winching : 2
directionalfelling : 2
specieslax : FALSE
objectivelax : TRUE

iter = 2
cores = 2
}
\keyword{datasets}
