% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingSimulationOutputs_iter.R
\docType{data}
\name{LoggingSimulationOutputs_iter}
\alias{LoggingSimulationOutputs_iter}
\title{Logging Simulation Outputs with iteration and parallelization}
\format{
list
}
\usage{
LoggingSimulationOutputs_iter
}
\description{
Outputs of the logging simulation on the plot 6 of Paracou (French Guiana)
}
\details{
With the parameters:
scenario : manual
objective : 10 m3/ha
fuel : 2
diversification : TRUE
winching : 2
directionalfelling : 2
specieslax : FALSE
objectivelax : TRUE

iter = 2
cores = 2
}
\keyword{datasets}
