% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingsummary.R
\name{loggingsummary}
\alias{loggingsummary}
\title{Summary of the outputs of the loggingsimulation function}
\usage{
loggingsummary(x)
}
\arguments{
\item{x}{Outputs of the loggingsimulation function (list)}
}
\value{
A summary of the outputs of the loggingsimulation function:
\itemize{
\item a reminder of the inputs
\item the outgoing figures
\item a print of the output inventory
}
}
\description{
Summary of the outputs of the loggingsimulation function
}
\examples{
data(Paracou6_2016) # inventory
data(DTMParacou) # topography
data(VerticalCreekHeight) # relative elevation
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters
data(ParamCrownDiameterAllometry) # parameters values of the crown diameter allometry
data(HarvestablePolygons)
data(MainTrails)

Rslt <- suppressMessages(
  loggingsimulation(Paracou6_2016, topography = DTMParacou,
                    relativeelevation  = VerticalCreekHeight, speciescriteria = SpeciesCriteria,
                    volumeparameters = ForestZoneVolumeParametersTable, scenario = "manual",
                    objective = 20, fuel = "2", diversification = TRUE, winching = "2",
                    directionalfelling = "2", specieslax = FALSE, objectivelax = TRUE,
                    maintrails = MainTrails,
                    crowndiameterparameters = ParamCrownDiameterAllometry,
                    harvestablepolygons = HarvestablePolygons,
                    advancedloggingparameters = loggingparameters(), iter = 1, cores = 1)
)

loggingsummary(Rslt)

}
