% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maintrailextract.R
\name{maintrailextract}
\alias{maintrailextract}
\title{Generating maintrails function}
\usage{
maintrailextract(topography, advancedloggingparameters = loggingparameters())
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer \strong{with a crs})
We advise you to generate your raster with Qgis rather than with the
'raster' package on R.}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
The maintrails outside of the topography raster (sf LINESTRING object
\strong{without crs})
}
\description{
Generating maintrails function
}
\examples{
data(DTMParacou)
MainTrails <- maintrailextract(DTMParacou)

library(raster)
plot(DTMParacou)
plot(MainTrails, add = TRUE)

}
