% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maintrails.R
\docType{data}
\name{MainTrails}
\alias{MainTrails}
\title{Maintrails}
\format{
A data.frame (sfc_LINESTRING)
}
\usage{
MainTrails
}
\description{
Maintrails lines defined at the extent of plot 6 of Paracou (French Guiana)
}
\keyword{datasets}
