% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONFGuyafortaxojoin.R
\name{ONFGuyafortaxojoin}
\alias{ONFGuyafortaxojoin}
\title{ONFGuyafortaxojoin}
\usage{
ONFGuyafortaxojoin(inventory, speciescriteria = SpeciesCriteria)
}
\arguments{
\item{inventory}{(data.frame)}

\item{speciescriteria}{mettre le lien vers le doc de la table (data.frame)}
}
\value{
The forest inventory with economic informations (economic name, rank, mimimum/maximum falling diameter, upped mimimum falling diameter)
per individuals (data.frame)
}
\description{
Joins a commercial criteria table to a forest inventory
}
\examples{
data(Paracou6_2016)
ONFGuyafortaxojoin(Paracou6_2016, SpeciesCriteria)

}
