% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Paracou6_2016.R
\docType{data}
\name{Paracou6_2016}
\alias{Paracou6_2016}
\title{Paracou6_2016}
\format{
A tibble with 3620 rows and 31 variables:
\describe{
Forest "character"
\item{var1}{description, in units}
\item{var2}{description, in units}
...
}
}
\source{
\url{http://paracou.cirad.fr}

EcoFoG::Guyafor2df(WHERE = "Forest='Paracou' AND Plot='6' AND CensusYear=2016",UID = NULL,PWD = NULL,Driver = "SQL Server Native Client 10.0")
}
\usage{
Paracou6_2016
}
\description{
Species inventory
2016 inventory of Paracou plot 6.Public dataset extracted from the Guyafor database.
}
\keyword{datasets}
