% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTopo.R
\docType{data}
\name{PlotTopo}
\alias{PlotTopo}
\title{Topography of Paracou plot 6}
\format{
RasterLayer
}
\usage{
PlotTopo
}
\description{
Cutting out the topography by the plot's boundaries (plot 6 of Paracou (1m
resolution, LiDAR campaign of 2016) (French Guiana)
}
\keyword{datasets}
