% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenariosparameters.R
\name{scenariosparameters}
\alias{scenariosparameters}
\title{scenariosparameters}
\usage{
scenariosparameters(
  scenario,
  objective = NULL,
  fuel = NULL,
  diversification = NULL,
  winching = NULL,
  directionalfelling = NULL
)
}
\arguments{
\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{objective}{Objective volume per hectare (numeric)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", damages and purge
exploitation in fuelwood = "1", exploitation of hollow trees, damages and purge in
fuelwood = "2"}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{winching}{No cable or grapple = "0", only cable = "1", grapple + cable =
"2"}

\item{directionalfelling}{Directional felling = "0" (absent), "1" (only to
avoid damage to future and reserve trees), "2" (avoid damage to future and
reserve trees + track orientation)}
}
\value{
A named list of 5 objects.
}
\description{
scenariosparameters
}
\examples{
scenariosparameters(scenario = "RIL1")

}
