% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondaryTrails.R
\docType{data}
\name{SecondaryTrails}
\alias{SecondaryTrails}
\title{Secondary trails of Paracou plot 6}
\format{
list
}
\usage{
SecondaryTrails
}
\description{
A list with :
}
\details{
\itemize{
\item RawSecondTrails : non-smoothed secondary trails
\item TrailsIdentity: information on sections of the trails with:
\itemize{
\item LineID:
\item LoggedTrees: idTree of trees reached by the trails
\item TypeExpl: type of winching
}
\item SmoothedTrails: Smoothed main and secondary trails polygons
(sfc_MULTIPOLYGON)
\item MainTrailsAccess : Random access point of maintrail for each PU
(sfc_MULTIPOLYGON)
\item TrailsDensity: Second trails density (in m/ha)
\item inventory: Updated inventory
\item CostRasterAgg: A cost Raster (RasterLayer)
}
}
\keyword{datasets}
