% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secondtrailsopening.R
\name{secondtrailsopening}
\alias{secondtrailsopening}
\title{secondtrailsopening}
\usage{
secondtrailsopening(
  inventory,
  DEM = DEMParacou,
  plotslope = PlotSlope,
  plots = Plots,
  prospectionunit = ProspectionUnit,
  maintrails = MainTrail,
  selectedtreespoints = SelectedTreesPoints,
  hollowtreespoints = HollowTreesPoints,
  avenirtreespoints = AvenirTreesPoint,
  reservetreespoints = ReserveTreesPoints,
  deadtreespoints = DeadTreesPoints,
  otherlogginparameters = logginparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{DEM}{(raster)}

\item{plotslope}{(raster)}

\item{otherlogginparameters}{(list)}
}
\value{

}
\description{
secondtrailsopening
}
\examples{

}
