% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothtrails.R
\name{smoothtrails}
\alias{smoothtrails}
\title{Trails smoothing}
\usage{
smoothtrails(
  paths,
  plotmask,
  verbose = FALSE,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{paths}{Raw secondary trails polygons (sp polylines)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{verbose}{return message on second trails density criteria (boolean)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with :
\enumerate{
\item SmoothedTrails: Smoothed trails polygons;
\item Smoothed second trails polygons;
\item Smoothed main trails polygons;
\item smoothed trail density.
}
}
\description{
Trails smoothing
}
\examples{
\dontrun{
data(SecondaryTrails)
SecondTrailsSmth <- smoothtrails(paths = SecondaryTrails$RawSecondTrails,
                                 plotmask = PlotMask,
                                   )
        }

}
