% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_temporaryintersection.R
\name{st_temporaryintersection}
\alias{st_temporaryintersection}
\title{st_temporaryintersection}
\usage{
st_temporaryintersection(
  topography,
  plotmask,
  treebuffers,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{treebuffers}{Polygons vector}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
sf_Polygons
}
\description{
st_temporaryintersection
}
\examples{

}
