% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefelling.R
\name{successfail}
\alias{successfail}
\title{successfail}
\usage{
successfail(
  inventory,
  fuel,
  directionalfelling,
  advancedloggingparameters = loggingparameters()
)

successfail(
  inventory,
  fuel,
  directionalfelling,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", damage
exploitation in fuelwood = "1", exploitation of hollow trees and damage in
fuelwood = "2"}

\item{directionalfelling}{Directional felling = "0" (absent), "1" (only to
avoid damage to future and reserve trees), "2" (avoid damage to future and
reserve trees + track orientation)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
Your inventory with: "TreeFellingOrientationSuccess" new column (see
the outputs metadata in the \code{\link{vignette}}).

Your inventory with: "TreeFellingOrientationSuccess" new column (see
the outputs metadata in the \code{\link{vignette}}).
}
\description{
Defines whether the directed fall of the tree is successful or
not by drawing in a Bernoulli distribution where the probability of success
is by default 60\%, and can be changed with the
\code{advancedloggingparameters} argument.

Defines whether the directed fall of the tree is successful or
not by drawing in a Bernoulli distribution where the probability of success
is by default 60\%, and can be changed with the
\code{advancedloggingparameters} argument.
}
\examples{
inventory <- addtreedim(inventorycheckformat(Paracou6_2016))
inventory <- treeselection(inventory, objective = 20, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 objectivelax = FALSE, DEM = DemParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory

successfail(inventory,fuel = "2",directionalfelling = "2",
advancedloggingparameters = loggingparameters())

inventory <- addtreedim(inventorycheckformat(Paracou6_2016))
inventory <- treeselection(inventory, objective = 20, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 objectivelax = FALSE, DEM = DemParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory

successfail(inventory,fuel = "2",directionalfelling = "2",
advancedloggingparameters = loggingparameters())

}
