% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createcanopy.R
\name{treefromthesky}
\alias{treefromthesky}
\title{treefromthesky}
\usage{
treefromthesky(dat)
}
\arguments{
\item{dat}{1 row data.frame with columns:
Xutm, Yutm (Tree coordinates), CrownDiameter, CrownHeight}
}
\value{
an ellipse (sfc_POLYGON) as a crown, with its diameter and height
filled in, representing the tree from the sky.
}
\description{
treefromthesky
}
\examples{
data(Paracou6_2016)
inventory <- addtreedim(inventorycheckformat(Paracou6_2016))
dat <- inventory[679,]

Crown <- treefromthesky(dat)

library(ggplot2)
ggplot() +
 geom_sf(data = sf::st_as_sf(inventory, coords = c("Xutm", "Yutm"))) +
 geom_sf(data = Crown, fill = "forestgreen") # trees polygons

}
