% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeselection.R
\name{treeselection}
\alias{treeselection}
\title{treeselection}
\usage{
treeselection(
  inventory,
  objective,
  type = "manual",
  fuel,
  diversification,
  specieslax = FALSE,
  objectivelax = FALSE,
  DEM = DemParacou,
  plotslope = PlotSlope,
  speciescriteria = SpeciesCriteria,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{objective}{Objective volume per hectare (numeric)}

\item{type}{"RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character)}

\item{fuel}{no  exploitation = "0", damage exploitation in fuelwood = "1",
exploitation of hollow trees and damage in fuelwood = "2"}

\item{diversification}{taking of other species in addition to the main commercial species (logical)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the objective volume (if stand too poor), = FALSE by default (logical)}

\item{DEM}{(RasterLayer)}

\item{plotslope}{(RasterLayer)}

\item{speciescriteria}{(data.frame)}

\item{otherloggingparameters}{(list) MainTrail (multiline)}
}
\value{

}
\description{
treeselection
}
\examples{
inventory <- addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))

treeselectionoutputs <- treeselection(inventory, SpeciesCriteria,
type ="manual", fuel = "2",objective = 20, diversification = FALSE, specieslax = FALSE,
objectivelax = FALSE, DEM = DemParacou, plotslope = PlotSlope, otherloggingparameters = loggingparameters()) # , MainTrail

save(treeselectionoutputs, file = "C:/Users/Utilisateur/AppData/Local/ProjetsR/Maria/treeselectionoutputs.Rdata")

}
