% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerticalCreekHeight.R
\docType{data}
\name{VerticalCreekHeight}
\alias{VerticalCreekHeight}
\title{Vertical creek height}
\format{
A Large RasterLayer
}
\usage{
VerticalCreekHeight
}
\description{
Vertical creek height of plot 6 of Paracou (French Guiana) (1 m resolution, 2016)
}
\keyword{datasets}
