% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VolumeParametersTable.R
\docType{data}
\name{VolumeParametersTable}
\alias{VolumeParametersTable}
\title{VolumeParametersTable}
\format{
A tibble with 4 rows and 3 variables: \describe{
\item{Zone}{Geographic zone (character)} \item{aCoef}{coefficient of the
equation: Volume = aCoef + bCoef * DBH^2 (numeric)}
\item{bCoef}{coefficient of the equation: Volume = aCoef + bCoef * DBH^2
(numeric)}
... }
}
\usage{
VolumeParametersTable
}
\description{
Volume parameters table to compute the harvestable volume of each tree,
depend to its geographic zone
}
\keyword{datasets}
