% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtreedim.R
\name{addtreedim}
\alias{addtreedim}
\title{Compute tree dimensions (Height, diameter, harvestable volume, including
crown)}
\usage{
addtreedim(
  inventory,
  crowndiameterparameters = ParamCrownDiameterAllometry,
  volumeparameters = ForestZoneVolumeParametersTable,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{crowndiameterparameters}{Crown diameter allometry parameters table to
compute the crown diameter of each tree, depend to its DBH (Diameter at
Breast Height) and its species, genus or family (data.frame)}

\item{volumeparameters}{Volume parameters table to compute the harvestable
volume of each tree, depend to its geographic zone if several locations
(data.frame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
(list)}
}
\value{
inventory (data.frame) with additional variables (TreeHeight,
TreeHarvestableVolume, TrunkHeight, CrownHeight, CrownDiameter)
}
\description{
Compute tree dimensions (Height, diameter, harvestable volume, including
crown)
}
\examples{
data(Paracou6_2016)
data(ParamCrownDiameterAllometry)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment

if (!("DBH" \%in\% names(Paracou6_2016))) {
tibble::add_column(Paracou6_2016, DBH = NA) #if DBH doesn't exist create it
Paracou6_2016$DBH = Paracou6_2016$CircCorr/pi
} # and compute it
Paracou6_2016 <- dplyr::filter(Paracou6_2016, DBH >= 10)

addtreedim(inventory = Paracou6_2016)

}
