% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtreedim.R
\name{addtreedim}
\alias{addtreedim}
\title{Compute tree dimensions (Height, diameter, harvestable volume, including crown)}
\usage{
addtreedim(
  inventory,
  crowndiameterparameters = ParamCrownDiameterAllometry,
  volumeparameters = ForestZoneVolumeParametersTable,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{crowndiameterparameters}{(data.frame)}

\item{volumeparameters}{(data.frame)}

\item{otherloggingparameters}{(list)}
}
\value{
inventory (data.frame) with additional variables (TreeHeight, TreeHarvestableVolume, TrunkHeight, CrownHeight, CrownDiameter)
}
\description{
Compute tree dimensions (Height, diameter, harvestable volume, including crown)
}
\examples{
data(Paracou6_2016)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
load(system.file("extdata", "ParamCrownDiameterAllometry.rda", package = "Maria"))

if (!("DBH" \%in\% names(Paracou6_2016))) {tibble::add_column(Paracou6_2016, DBH = NA) #if DBH doesn't exist create it
Paracou6_2016$DBH = Paracou6_2016$CircCorr/pi} # and compute it

addtreedim(inventory = Paracou6_2016)
}
