% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addtreedim.R
\name{addtreedim}
\alias{addtreedim}
\title{Compute tree dimensions (Height, diameter, harvestable volume, including
crown)}
\usage{
addtreedim(
  inventory,
  volumeparameters,
  crowndiameterparameters = ParamCrownDiameterAllometry,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{volumeparameters}{Volume parameters table (in the same format of
\code{\link{ForestZoneVolumeParametersTable}}) to compute the harvestable
volume of each tree, depend to its geographic zone if several locations
(data.frame)}

\item{crowndiameterparameters}{Crown diameter allometry parameters table (in
the same format of \code{\link{ParamCrownDiameterAllometry}}) to compute
the crown diameter of each tree, depend to its DBH (Diameter at Breast
Height) and its species, genus or family. (data.frame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
inventory (data.frame) with additional variables (TreeHeight,
TreeHarvestableVolume, TrunkHeight, CrownHeight, CrownDiameter)
}
\description{
Compute tree dimensions (Height, diameter, harvestable volume, including
crown)
}
\examples{
data(Paracou6_2016)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
data(ParamCrownDiameterAllometry)

if (!("DBH" \%in\% names(Paracou6_2016))) {
tibble::add_column(Paracou6_2016, DBH = NA) #if DBH doesn't exist create it
Paracou6_2016$DBH = Paracou6_2016$CircCorr/pi
} # and compute it
Paracou6_2016 <- dplyr::filter(Paracou6_2016, DBH >= 10)

addtreedim(Paracou6_2016, volumeparameters = ForestZoneVolumeParametersTable)

}
