% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{Say "Bonjour !" with a name}
\usage{
bonjour(nom = "Maria", verbose = TRUE)
}
\arguments{
\item{nom}{(character) Name of the person to say bonjour.}

\item{verbose}{(boolean) Print in console}
}
\value{

}
\description{
Say "Bonjour !" with a name
}
\examples{
bonjour()
}
