% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bonjour.R
\name{bonjour}
\alias{bonjour}
\title{Say "Bonjour !" with a name}
\usage{
bonjour(nom = "Maria", verbose = TRUE)
}
\arguments{
\item{nom}{Name of the person to say "Bonjour !" (character).}

\item{verbose}{Print in console (boolean).}
}
\value{
A character string "Bonjour! Name"
}
\description{
Say "Bonjour !" with a name
}
\examples{
bonjour()
}
