% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaninventory.R
\name{cleaninventory}
\alias{cleaninventory}
\title{Clean inventory (only alive trees within the inventoried plot)}
\usage{
cleaninventory(
  inventory,
  plotmask,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
vignette) (data.frame)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
The inventory (data.frame) with only alive trees within the
inventoried plot ('TreesIn' = "1").
}
\description{
Clean inventory (only alive trees within the inventoried plot)
}
\examples{
data(Paracou6_2016) # inventory
data(PlotMask) # the inventoried plot mask
new <- cleaninventory(Paracou6_2016,
                      PlotMask, loggingparameters(MinDBHValue = 5))

}
