% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaninventory.R
\name{cleaninventory}
\alias{cleaninventory}
\title{Select trees according to the protocol}
\usage{
cleaninventory(inventory)
}
\arguments{
\item{inventory}{Your inventory data or the "Paracou6_2016" test data (data.frame)}
}
\value{
inventory (data.frame).Stop the function if the tree identifiers (idTree) are not unique, or if there are different plots in the inventory.
}
\description{
Select trees according to the protocol: alive, unique, at the right place & of the required minimum protocol diameter
}
\examples{

data(Paracou6_2016)
cleaninventory(inventory = Paracou6_2016)

}
