% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createcanopy.R
\name{createcanopy}
\alias{createcanopy}
\title{createcanopy}
\usage{
createcanopy(inventory)
}
\arguments{
\item{inventory}{Your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}
}
\value{
a dataframe with a column 'Crowns' containing the ellipses
(sfc_POLYGON) as trees crown, with their diameter and height filled in,
representing trees from the sky.
}
\description{
createcanopy
}
\examples{
data(Paracou6_2016)
Paracou6_2016 <- dplyr::slice(Paracou6_2016, 1:10)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- createcanopy(inventory)

# The small ones first so that they are behind the big ones on the plot
inventory <- dplyr::arrange(inventory, TreeHeight)
library(ggplot2)
ggplot() +
  geom_sf(data = getgeometry(inventory, Crowns),
          aes(alpha = TreeHeight),
          fill = "forestgreen")

}
