% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefelling.R
\name{directionalfellingsuccessdef}
\alias{directionalfellingsuccessdef}
\title{Directional felling success definition}
\usage{
directionalfellingsuccessdef(
  inventory,
  fuel,
  directionalfelling,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"}

\item{directionalfelling}{Directional felling = "0" (absent), "1" (only to
avoid damage to future and reserve trees), "2" (to avoid damage to future and
reserve trees and to position the log relative to the track)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
Input inventory with the "TreeFellingOrientationSuccess" new column (see
the outputs metadata in the \code{\link{vignette}}).
}
\description{
Defines whether the directed fall of the tree is successful or
not by drawing in a Bernoulli distribution where the probability of success
is by default 60\%, and can be changed with the
\code{advancedloggingparameters} argument.
}
\examples{
inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- treeselection(inventory, objective = 20, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 objectivelax = TRUE, topography = DTMParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory

new <- directionalfellingsuccessdef(inventory,fuel = "2", directionalfelling = "2",
advancedloggingparameters = loggingparameters())

}
