% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMParacou.R
\docType{data}
\name{DTMParacou}
\alias{DTMParacou}
\title{Paracou Digital terrain model (DTM)}
\format{
A RasterLayer
}
\usage{
DTMParacou
}
\description{
Digital terrain model (DTM) of plot 6 of Paracou (French Guiana) (1m resolution, 2016)
}
\keyword{datasets}
