% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DTMParacou.R
\docType{data}
\name{DTMParacou}
\alias{DTMParacou}
\title{Paracou Digital terrain model (DTM)}
\format{
A Large RasterLayer
}
\source{
This study benefited from funding by the Centre National
d’Etudes Spatiales (CNES Tosca Leaf EXPEVAL) who provided the ALS data used
to compute the DTM
}
\usage{
DTMParacou
}
\description{
Digital terrain model (DTM) of plot 6 of Paracou (French Guiana) (1 m
resolution, 2016)
}
\keyword{datasets}
