% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitablefuelwoodvolume.R
\name{exploitablefuelwoodvolume}
\alias{exploitablefuelwoodvolume}
\title{exploitablefuelwoodvolume}
\usage{
exploitablefuelwoodvolume(
  inventory,
  fuel,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{fuel}{no  exploitation = "0", damage exploitation in fuelwood = "1",
exploitation of hollow trees and damage in fuelwood = "2"}

\item{advancedloggingparameters}{(list)}
}
\value{
A list with the damage volume, and the fuelwood volume when fuelwood exploitation is chosen.
}
\description{
exploitablefuelwoodvolume
}
\examples{

\dontrun{
data(Paracou6_2016)
inventory <- addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))

inventory <- treeselection(inventory, SpeciesCriteria,
scenario ="manual", fuel = "2", objective = 20, diversification = TRUE, specieslax = FALSE,
objectivelax = FALSE, advancedloggingparameters = loggingparameters())$inventory

inventory <- secondtrailsopening()$inventory

inventory <- treefelling()$inventory

inventory <- adjustedsecondtrails()$inventory

inventory <- timberharvestedvolume()$inventory

exploitablefuelwoodvolume(inventory, fuel = "2", advancedloggingparameters = loggingparameters())
}

}
