% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploitablefuelwoodvolume.R
\name{exploitablefuelwoodvolume}
\alias{exploitablefuelwoodvolume}
\title{exploitablefuelwoodvolume}
\usage{
exploitablefuelwoodvolume(
  inventory,
  scenario,
  fuel,
  advancedloggingparameters = loggingparameters(),
  LoggedVolume = LoggedVolume,
  NoHollowLoggedVolume = NoHollowLoggedVolume
)
}
\arguments{
\item{inventory}{Your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", damage
exploitation in fuelwood = "1", exploitation of hollow trees and damage in
fuelwood = "2"}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}

\item{LoggedVolume}{All the logged volume (in m^3) (numeric)}

\item{NoHollowLoggedVolume}{The healthy logged volume (in m^3) (numeric)}
}
\value{
A list with the damage volume, and the fuelwood volume when fuelwood
exploitation is chosen.
}
\description{
exploitablefuelwoodvolume
}
\examples{
data(Paracou6_2016)

MainTrail <- sf::st_linestring(matrix(c(286400, 583130,
                                        286400, 583250,
                                        286655, 583250,
                                        286655, 583130,
                                        286400, 583130) # the return
                                     ,ncol=2, byrow=TRUE))

pol1 <- list(matrix(c(286503, 583134,
                      286503, 583240,
                      286507, 583240,
                      286507, 583134,
                      286503, 583134) # the return
                   ,ncol=2, byrow=TRUE))
pol2 <- list(matrix(c(286650, 583134,
                      286650, 583240,
                      286654, 583240,
                      286654, 583134,
                      286650, 583134) # the return
                   ,ncol=2, byrow=TRUE))

PolList = list(pol1,pol2) #list of lists of numeric matrices
ScndTrail <- sf::st_multipolygon(PolList)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- suppressMessages(treeselection(inventory, objective = 30, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 objectivelax = FALSE, topography = DTMParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory)

inventory <- treefelling(inventory, scenario = "manual", fuel = "2",
directionalfelling = "2", MainTrail = MainTrail, ScndTrail = ScndTrail,
advancedloggingparameters = loggingparameters())

TimberV <- timberharvestedvolume(inventory, scenario = "manual", fuel = "2",
advancedloggingparameters = loggingparameters())

LoggedVolume <- TimberV$LoggedVolume
NoHollowLoggedVolume <- TimberV$NoHollowLoggedVolume

exploitablefuelwoodvolume(inventory, scenario = "manual", fuel = "2",
advancedloggingparameters = loggingparameters(),
LoggedVolume = LoggedVolume, NoHollowLoggedVolume = NoHollowLoggedVolume)

}
