% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAccesExplArea.R
\name{filteraccesexplarea}
\alias{filteraccesexplarea}
\title{Filter access to harvestable zone}
\usage{
filteraccesexplarea(
  harvestablepolygons,
  MainTrails,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{harvestablepolygons}{Accessible area of the inventoried plot (default:
\code{\link{harvestableareadefinition}}) (sf polygons data.frame)}

\item{MainTrails}{Main trails defined at the entire harvestable area (sf
polylines)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
Accessible area of the inventoried plot according to selected
logging scenario (sf polygons data.frame)
}
\description{
Filter access to harvestable zone
}
\examples{
data(MainTrails)
data(HarvestablePolygons)

AccessPolygones <- filteraccesexplarea(
 harvestablepolygons = HarvestablePolygons,
 MainTrails = MainTrails,
 advancedloggingparameters = loggingparameters())


}
