% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterAccesExplArea.R
\name{FilterAccesExplArea}
\alias{FilterAccesExplArea}
\title{FilterAccesExplArea}
\usage{
FilterAccesExplArea(
  harvestablepolygons,
  maintrails,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{harvestablepolygons}{Accessible area of the inventoried plot
(default: \code{\link{HarvestableAreaDefinition}}) (sf polygons data.frame)}

\item{maintrails}{Main trails defined at the entire harvestable area (sf polylines)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
Accessible area  of the inventoried plot according to selected logging scenario (sf polygons data.frame)
}
\description{
FilterAccesExplArea
}
\examples{

 AccessPolygones <-  FilterAccesExplArea (harvestablepolygons = HarvestablePolygons,
 maintrails = MainTrails,
 advancedloggingparameters = loggingparameters())

}
