% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurereserve.R
\name{futurereserve}
\alias{futurereserve}
\title{Future & reserve trees designation}
\usage{
futurereserve(
  inventory,
  speciescriteria,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
vignette) (data.frame)}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in
the same format as \code{\link{SpeciesCriteria}} (data.frame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
Input inventory with selected future and reserve trees
(\emph{LoggingStatus} = "\emph{future}"/"\emph{reserve}"),
for your logging criteria (\code{\link{SpeciesCriteria}})
}
\description{
designates which trees will be harvested in future seasons:
"future trees", and which should be protected as seed trees: "reserve
trees".
}
\details{
\strong{Future} trees are all trees satisfying the following conditions:
\itemize{
\item species of 1st economic rank
\item DBH between 35cm ('FutureTreesMinDiameter') and the species \emph{MinFD}
or \emph{UpMinFD} if it has been raised for its species.
\item in the harvestable zones
}

\strong{Reserve} trees are randomly chosen among future trees so that
the number of reserve trees is equal to the number of harvested trees.
}
\examples{
data(Paracou6_2016)
data(ParamCrownDiameterAllometry)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
data(SpeciesCriteria)
data(DTMParacou)
data(MainTrails)
data(HarvestableAreaOutputsCable)

scenario <- "RIL2"

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory,
 topography = DTMParacou,
 diversification = TRUE,
 specieslax = FALSE,
 plotslope = HarvestableAreaOutputsCable$PlotSlope,
 maintrails = MainTrails,
 harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
 scenario = scenario,
 advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, scenario = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = TRUE,
topography = DTMParacou,
advancedloggingparameters = loggingparameters(), VO = 125,
 HVinit = HVinit)$inventory

futurereserve(inventory, SpeciesCriteria)

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{loggingparameters}}
}
