% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurereserve.R
\name{futurereserve}
\alias{futurereserve}
\title{futurereserve}
\usage{
futurereserve(
  inventory,
  speciescriteria = SpeciesCriteria,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{speciescriteria}{(data.frame)}

\item{otherloggingparameters}{(list)}
}
\value{

}
\description{
futurereserve
}
\examples{
inventory <- harvestable(ONFGuyafortaxojoin(addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))),
diversification = TRUE)$inventory

HVinit <- harvestable(ONFGuyafortaxojoin(addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))),
diversification = TRUE)$HVinit

inventory <- selected(inventory, type = "manual", fuel = "0", diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters(), VO = 80, HVinit = HVinit)$inventory

futurereserve(inventory)

}
