% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurereserve.R
\name{futurereserve}
\alias{futurereserve}
\title{futurereserve}
\usage{
futurereserve(
  inventory,
  speciescriteria = SpeciesCriteria,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{speciescriteria}{(data.frame)}

\item{otherloggingparameters}{(list)}
}
\value{
A FAIRE
}
\description{
futurereserve
}
\examples{

\dontrun{
harvestableOutputs <- harvestable(ONFGuyafortaxojoin(inventory,
speciescriteria = speciescriteria),
diversification = diversification, specieslax = specieslax,
DEM = DemParacou, plotslope = PlotSlope,
otherloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, type = "manual", fuel = "0",
 diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters(), VO = 80,
 HVinit = HVinit)$inventory

futurereserve(inventory)
}

}
