% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurereserve.R
\name{futurereserve}
\alias{futurereserve}
\title{futurereserve}
\usage{
futurereserve(
  inventory,
  speciescriteria = SpeciesCriteria,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{speciescriteria}{(data.frame)}

\item{otherloggingparameters}{(list)}
}
\value{
A FAIRE
}
\description{
futurereserve
}
\examples{

data(Paracou6_2016)
data(DemParacou)

inventory <- ONFGuyafortaxojoin(addtreedim(cleaninventory(
inventorycheckformat(Paracou6_2016))))

harvestableOutputs <- harvestable(inventory, diversification = TRUE,
 specieslax = FALSE,
DEM = DemParacou, plotslope = PlotSlope,
otherloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, type = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters(), VO = 30,
 HVinit = HVinit)$inventory

futurereserve(inventory)

}
