% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/futurereserve.R
\name{futurereserve}
\alias{futurereserve}
\title{Future & reserve trees designation}
\usage{
futurereserve(
  inventory,
  speciescriteria,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in
the same format of \code{\link{SpeciesCriteria}} (data.frame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
your inventory with selected future and reserve trees, for your
logging criteria (\code{\link{SpeciesCriteria}})
}
\description{
Future & reserve trees designation
}
\examples{
data(Paracou6_2016)
data(ParamCrownDiameterAllometry)
data(ForestZoneVolumeParametersTable) # The volume parameters data in the global environment
data(SpeciesCriteria)
data(DTMParacou)
data(PlotSlope)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- ONFGuyafortaxojoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, diversification = TRUE,
 specieslax = FALSE,
topography = DTMParacou, plotslope = PlotSlope,
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

inventory <- selected(inventory, scenario = "manual", fuel = "0",
diversification = TRUE, specieslax = FALSE, objectivelax = TRUE,
topography = DTMParacou,
advancedloggingparameters = loggingparameters(), VO = 125,
 HVinit = HVinit)$inventory

futurereserve(inventory, SpeciesCriteria)

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{loggingparameters}}
}
