% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestable.R
\name{harvestable}
\alias{harvestable}
\title{Harvestable trees identification}
\usage{
harvestable(
  inventory,
  topography,
  diversification,
  specieslax = FALSE,
  harvestablepolygons,
  plotslope,
  scenario,
  winching = NULL,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)
#'}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by
default (logical)}

\item{harvestablepolygons}{Accessible area of the inventoried plot
(default: \code{\link{harvestableareadefinition}}) (sfc_MULTIPOLYGON)}

\item{plotslope}{Slopes (in radians) of the inventoried plot (with a
neighbourhood of 8 cells) (default: \code{\link{PlotSlope}}) (RasterLayer)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{winching}{"0": no cable or grapple (trail to tree foot)
"1": only cable (default = 40m)
"2": grapple (default = 6m) + cable (grapple priority)
If grapple + cable (winching = "2") without fuel wood (fuel = "0")
recovery of the tree foot with grapple if possible (respected grapple
conditions) otherwise with cable with angle to the trail.
Avoidance of future/reserves if chosen.}

\item{advancedloggingparameters}{Other parameters of the logging simulator
(\code{\link{loggingparameters}}) (list)}
}
\value{
input inventory with new columns:
\itemize{
\item The exploitability criteria ("DistCriteria", "Slope"(in radians), "SlopeCriteria"), and if
they are validated for each of the trees ("LoggingStatus").
\item The probability of a tree having visible defects ("VisibleDefectProba")
and the visible defect trees ("VisibleDefect").
}

The function returns the harvestable volume too, in the
plot for these criteria.
}
\description{
The function tells you the harvestable volume in the plot, and
which trees are harvestable according to your harvestability criteria
}
\details{
Trees will be designated as "\strong{harvestable}" if they:
\itemize{
\item belonging to species of 1st economic rank or more if diversification
\item DBH between the MinFD and the MaxFD.
\item not isolated ( >100m ('IsolateTreeMinDistance' in
\code{\link{loggingparameters}})) from other individuals of the same
species in the aggregative species case (\code{\link{SpeciesCriteria}},
'Aggregative' column).
\item on slopes < 22\% ('TreeMaxSlope'in \code{\link{loggingparameters}})
\item off the main trails.
}

Trees with visible defects are identified ('VisiblyDefectModel' in
'advancedloggingparameters' argument) among the trees with harvestable
criteria and are therefore considered 'non-harvestable'.
}
\examples{
data(Paracou6_2016)
data(DTMParacou)
data(PlotMask)
data(SpeciesCriteria)
data(HarvestableAreaOutputsCable)

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- commercialcriteriajoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
diversification = TRUE, specieslax = FALSE,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
scenario = "manual", winching = "0",
advancedloggingparameters = loggingparameters())

new <- harvestableOutputs$inventory

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{DTMParacou}}, \code{\link{PlotSlope}},
\code{\link{loggingparameters}}
}
