% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestable.R
\name{harvestable}
\alias{harvestable}
\title{harvestable}
\usage{
harvestable(
  inventory,
  topography,
  diversification,
  winching,
  specieslax = FALSE,
  DTM = DTMParacou,
  exploitpolygones = ExploitPolygones,
  maintrails = MainTrails,
  plotslope = PlotSlope,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (default: \code{\link{DTMParacou}}) (RasterLayer)}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by
default (logical)}

\item{exploitpolygones}{Accessible area of the inventoried plot
(default: \code{\link{HarvestableAreaDefinition}}) (sf polygons data.frame)}

\item{maintrails}{Main trails defined at the entire harvestable area (sf polylines)}

\item{plotslope}{Slopes (in radians) of the inventoried plot (with a
neighbourhood of 8 cells) (default: \code{\link{PlotSlope}}) (RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
Your inventory with the exploitability criteria, and if they are
validated for each of the trees. The function returns the harvestable volume
too, in the plot for these criteria.
}
\description{
The function tells you the harvestable volume in the plot, and
which trees are harvestable according to your harvestability criteria
}
\examples{

data(Paracou6_2016)

data(DTMParacou)
data(PlotSlope)
data(ExploitPolygones)
data(SpeciesCriteria)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- ONFGuyafortaxojoin(inventory, SpeciesCriteria)

test <- harvestable(inventory, topography = DTMParacou,exploitpolygones = ExploitPolygones,
diversification = TRUE, specieslax = FALSE, plotslope = PlotSlope,maintrails = Maintrails,
advancedloggingparameters = loggingparameters())

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{DTMParacou}}, \code{\link{PlotSlope}},
\code{\link{loggingparameters}}
}
