% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestable.R
\name{harvestable}
\alias{harvestable}
\title{Harvestable trees identification}
\usage{
harvestable(
  inventory,
  topography,
  harvestablepolygons,
  maintrails,
  plotslope,
  diversification,
  specieslax = FALSE,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{harvestablepolygons}{Accessible area of the inventoried plot
(default: \code{\link{HarvestableAreaDefinition}}) (sf polygons data.frame)}

\item{maintrails}{Main trails defined at the entire harvestable area (sf polylines)}

\item{plotslope}{Slopes (in radians) of the inventoried plot (with a
neighbourhood of 8 cells) (default: \code{\link{PlotSlope}}) (RasterLayer)}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by
default (logical)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
(\code{\link{loggingparameters}}) (list)}
}
\value{
input inventory with new columns:
\itemize{
\item The exploitability criteria ("DistCrit", "Slope"(in radians), "SlopeCrit"), and if
they are validated for each of the trees ("LoggingStatus").
\item The probability of a tree having visible defects ("VisibleDefectProba")
and the visible defect trees ("VisibleDefect").
}

The function returns the harvestable volume too, in the
plot for these criteria.
}
\description{
The function tells you the harvestable volume in the plot, and
which trees are harvestable according to your harvestability criteria
}
\details{
Trees will be designated as "\strong{harvestable}" if they:
\itemize{
\item belonging to species of 1st economic rank or more if diversification
\item DBH between the MinFD and the MaxFD.
\item not isolated ( >100m ('IsolateTreeMinDistance' in
\code{\link{loggingparameters}})) from other individuals of the same
species in the aggregative species case (\code{\link{SpeciesCriteria}},
'Aggregative' column).
\item on slopes < 22\% ('TreeMaxSlope'in \code{\link{loggingparameters}})
\item off the main trails.
}

Trees with visible defects are identified ('VisiblyDefectModel' in
'advancedloggingparameters' argument) among the trees with harvestable
criteria and are therefore considered 'non-harvestable'.
}
\examples{

data(Paracou6_2016)
data(DTMParacou)
data(PlotSlope)
data(SpeciesCriteria)
data(HarvestablePolygons)
data(MainTrails)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- ONFGuyafortaxojoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
harvestablepolygons = HarvestablePolygons, maintrails = MainTrails,
plotslope = PlotSlope, diversification = TRUE, specieslax = FALSE,
advancedloggingparameters = loggingparameters())

new <- harvestableOutputs$inventory

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{DTMParacou}}, \code{\link{PlotSlope}},
\code{\link{loggingparameters}}
}
