% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestable.R
\name{harvestable}
\alias{harvestable}
\title{harvestable}
\usage{
harvestable(
  inventory,
  diversification,
  specieslax = FALSE,
  DEM = DemParacou,
  plotslope = PlotSlope,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{diversification}{(logical)}

\item{specieslax}{= FALSE by default (logical)}

\item{DEM}{(RasterLayer)}

\item{plotslope}{(RasterLayer)}

\item{otherloggingparameters}{(list)}
}
\value{
Your inventory with the exploitability criteria, and if they are validated for each of the trees.
The function returns the harvestable volume too, in the plot for these criteria.
}
\description{
The function tells you the harvestable volume in the plot,
and which trees are harvestable according to your harvestability criteria
}
\examples{

data(Paracou6_2016)
data(DemParacou)

inventory <- ONFGuyafortaxojoin(addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016))))
test <- harvestable(inventory, diversification = TRUE, specieslax = FALSE,
DEM = DemParacou, plotslope = PlotSlope,otherloggingparameters = loggingparameters())

}
