% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaDefinition.R
\name{HarvestableAreaDefinition}
\alias{HarvestableAreaDefinition}
\title{HarvestableAreaDefinition}
\usage{
HarvestableAreaDefinition(
  plot = Plots,
  DEM = DEMParacou,
  HCrique = Hcrique,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{plot}{Studied plots (sf data.frame)}

\item{DEM}{Digital elevation model (raster)}

\item{HCrique}{Relative elevation from nearest channel network (raster)}

\item{advancedloggingparameters}{Other parameters of the logging simulator}
}
\value{
A collection of polygones defined as 1 : harvestable area / 0 : non-harvestable area
}
\description{
HarvestableAreaDefinition
}
\examples{

ExploitPolygones <- HarvestableAreaDefinition(Plot = Plots,DEM = DEM,HCrique = HCrique$distance)

}
