% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaDefinition.R
\name{HarvestableAreaDefinition}
\alias{HarvestableAreaDefinition}
\title{Harvestable area definition}
\usage{
HarvestableAreaDefinition(
  topography,
  verticalcreekheight,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{verticalcreekheight}{Relative elevation from nearest channel network
(Large RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with:
\itemize{
\item 'HarvestablePolygons': a collection of polygons (sf polygon) defined as:
1 : harvestable area,
0 : non-harvestable area
\item 'PlotSlope': a raster with slope (in radians) characteristic of the studied
plot (Large RasterLayer)
}
}
\description{
Harvestable area definition
}
\examples{
\dontrun{
data(PlotMask)
data(DTMParacou)
data(VerticalCreekHeight)

HarvestableAreaOutputs <- HarvestableAreaDefinition(
  topography = DTMParacou,
  verticalcreekheight = VerticalCreekHeight,
  advancedloggingparameters = loggingparameters()
  )

plot(HarvestableAreaOutputs[[1]])
}

}
