% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvestableareadefinition.R
\name{harvestableareadefinition}
\alias{harvestableareadefinition}
\title{Harvestable area definition}
\usage{
harvestableareadefinition(
  topography,
  verticalcreekheight,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{verticalcreekheight}{Relative elevation from nearest channel network
(Large RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with:
\itemize{
\item 'HarvestablePolygons': a collection of polygons (sf (sfc_POLYGON)) defined as:
1 : harvestable area,
0 : non-harvestable area
\item 'PlotSlope': a raster with slope (in radians) characteristic of the studied
plot (Large RasterLayer)
\item HarvestableArea : the harvestable area in hectares (double).
}
}
\description{
Harvestable area definition
}
\examples{
\dontrun{
data(PlotMask)
data(DTMParacou)
data(VerticalCreekHeight)

HarvestableAreaOutputs <- harvestableareadefinition(
  topography = DTMParacou,
  verticalcreekheight = VerticalCreekHeight,
  advancedloggingparameters = loggingparameters()
  )

library(ggplot2)
library(sf)
ggplot() +
geom_sf(data = HarvestableAreaOutputs[[1]],
        aes(alpha = Harvestable),
        fill = "olivedrab") +
  labs(alpha = "Harvestable") +
  labs(title = "P6  Harvestable zones")

HarvestableAreaOutputs[[2]]
}

}
