% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaDefinition.R
\name{harvestableareadefinition}
\alias{harvestableareadefinition}
\title{Harvestable zones definition within the plot}
\usage{
harvestableareadefinition(
  topography,
  creekverticaldistance,
  creekhorizontaldistance,
  maintrails,
  plotmask,
  scenario,
  winching = NULL,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer \strong{with a crs in UTM})}

\item{creekverticaldistance}{Relative vertical distance
(1 m resolution) from nearest channel network
(RasterLayer \strong{with a crs in UTM}) (See \code{\link{CreekDistances}})
To generate creek distances: \code{\link{CreekDistances}} in 'Articles'.}

\item{creekhorizontaldistance}{Relative horizontal distance
(1 m resolution) from nearest channel network
(RasterLayer \strong{with a crs in UTM}) (See \code{\link{CreekDistances}})
To generate creek distances: \code{\link{CreekDistances}} in 'Articles'.}

\item{maintrails}{Main trails defined at the entire harvestable area (sf
linestring \strong{with a crs in UTM})}

\item{plotmask}{Inventoried plot mask
(SpatialPolygonsDataFrame \strong{with a crs in UTM})}

\item{scenario}{Logging scenario among: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
vignette)}

\item{winching}{"0": no cable or grapple (trail to tree foot)
"1": only cable (default = 40m)
"2": grapple (default = 6m) + cable (grapple priority)
If grapple + cable (winching = "2") without fuel wood (fuel = "0")
recovery of the tree foot with grapple if possible (respected grapple
conditions) otherwise with cable with angle to the trail.
Avoidance of future/reserves if chosen.}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with:
\itemize{
\item 'HarvestablePolygons': a collection of polygons (sfc_MULTIPOLYGON, with
crs) defined as:
1 : harvestable zone,
0 : non-harvestable zone
\item 'PlotSlope': Slopes of the plot (in radians) characteristic of the studied
plot (Large RasterLayer with crs)
\item HarvestableArea : the harvestable area in hectares (double)
\item MachinePolygons : a collection of polygons (sf (sfc_POLYGON with crs))
defined as:
1 : accessible machine zone,
0 : non-accessible machine zone
}
}
\description{
The simulator avoids the water sources buffer zones and slopes
inaccessible to machinery by taking the highest maximum slope percentage
tolerated by the machines.
}
\examples{
\dontrun{
data(PlotMask)
data(DTMParacou)
data(CreekDistances)
data(MainTrails)

HarvestableAreaOutputsCable <- harvestableareadefinition(
  topography = DTMParacou,
  creekverticaldistance = CreekDistances$distvert,
  creekhorizontaldistance = CreekDistances$disthorz,
  maintrails = MainTrails,
  plotmask = PlotMask,
  scenario = "manual", winching = "1",
  advancedloggingparameters = loggingparameters()
  )
}
data(HarvestableAreaOutputsCable)
library(ggplot2)
library(sf)
ggplot() +
  # Harvestable zones
  geom_sf(data = HarvestableAreaOutputsCable$HarvestablePolygons,
        fill = "olivedrab", alpha = 0.1) +
   geom_sf(data = HarvestableAreaOutputsCable$MachinePolygons,
        fill = "olivedrab", alpha = 0.5) +
  labs(alpha = "Harvestable") +
  labs(title = "Paracou P6 - Harvestable zones") +
  scale_colour_manual(values = c("Harvestable area" = "olivedrab"))

HarvestableAreaOutputsCable$PlotSlope

}
