% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaDefinition.R
\name{HarvestableAreaDefinition}
\alias{HarvestableAreaDefinition}
\title{HarvestableAreaDefinition}
\usage{
HarvestableAreaDefinition(
  plot,
  dtm,
  verticalcreekheight,
  advancedloggingparameters
)
}
\arguments{
\item{plot}{Studied plots (sf data.frame)}

\item{dtm}{Digital terrain model (raster)}

\item{verticalcreekheight}{Relative elevation from nearest channel network (raster)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A collection of polygones defined as 1 : harvestable area / 0 : non-harvestable area
}
\description{
HarvestableAreaDefinition
}
\examples{
data(Plots)
data(DTMParacou)
data(VerticalCreekHeight)


ExploitPolygones <- HarvestableAreaDefinition(plot = Plots,
                                             dtm = DTMParacou,
                                             verticalcreekheight = VerticalCreekHeight,
                                             advancedloggingparameters = loggingparameters())

}
