% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestableAreaDefinition.R
\name{HarvestableAreaDefinition}
\alias{HarvestableAreaDefinition}
\title{HarvestableAreaDefinition}
\usage{
HarvestableAreaDefinition(
  plot,
  dtm,
  verticalcreekheight,
  advancedloggingparameters
)
}
\arguments{
\item{plot}{Studied plot (SpatialPolygonsDataFrame)}

\item{dtm}{Digital terrain model (Large RasterLayer)}

\item{verticalcreekheight}{Relative elevation from nearest channel network
(Large RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A collection of polygons defined as 1 : harvestable area, 0 :
non-harvestable area
}
\description{
HarvestableAreaDefinition
}
\examples{
data(Plots)
data(DTMParacou)
data(VerticalCreekHeight)


HarvestableArea <- HarvestableAreaDefinition(plot = Plots,
                                             dtm = DTMParacou,
                                             verticalcreekheight = VerticalCreekHeight,
                                             advancedloggingparameters = loggingparameters())
plot(HarvestableArea)

}
