% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HarvestablePolygons.R
\docType{data}
\name{HarvestablePolygons}
\alias{HarvestablePolygons}
\title{HarvestablePolygons}
\format{
A data.frame (sfc_POLYGON)
}
\usage{
HarvestablePolygons
}
\description{
A collection of polygons defined as 1 : harvestable area / 0 : non-harvestable area
of plot 6 of Paracou (French Guiana)
}
\keyword{datasets}
