% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorycheckformat.R
\name{inventorycheckformat}
\alias{inventorycheckformat}
\title{Check & format your inventory data for the package "Maria"}
\usage{
inventorycheckformat(inventory)
}
\arguments{
\item{inventory}{(dataframe) Your inventory data or the "Paracou6_2016" test data#'}
}
\value{
stop the function if the format is not the one required.
}
\description{
Check & format your inventory data for the package "Maria"
}
\examples{

data(Paracou6_2016)
inventorycheckformat(inventory = Paracou6_2016)

}
