% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorycheckformat.R
\name{inventorycheckformat}
\alias{inventorycheckformat}
\title{Check & format your inventory data for the "Maria" package}
\usage{
inventorycheckformat(inventory)
}
\arguments{
\item{inventory}{Your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}
}
\value{
Stop the function if the format is not the one required.
}
\description{
Check & format your inventory data for the "Maria" package
}
\examples{

data(Paracou6_2016)
inventorycheckformat(inventory = Paracou6_2016)

}
