% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inventorycheckformat.R
\name{inventorycheckformat}
\alias{inventorycheckformat}
\title{Check the input inventory data for the "LoggingLab" package}
\usage{
inventorycheckformat(inventory)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}
}
\value{
The inventory if the inventory is in the required format.
Stop the function if the format is not the one required.
}
\description{
Check if the input inventory data is compatible with the "LoggingLab"
package (see 'Required format of the inventory' section of the vignette)
}
\examples{

data(Paracou6_2016)
inventorycheckformat(inventory = Paracou6_2016)

}
