% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoggingLab-package.R
\docType{package}
\name{LoggingLab}
\alias{LoggingLab}
\alias{LoggingLab-package}
\title{LoggingLab-package}
\description{
Logging simulator
}
\details{
An R Package to Simulate Forest Logging
}
\section{LoggingLab functions}{

loggingsimulation

loggingparameters

scenariosparameters

inventorycheckformat

addtreedim

treeselection

commercialcriteriajoin

harvestable

selected

futurereserve

treefelling

createcanopy

treefromthesk

directionalfellingsuccessdef

felling1tree

rotatepolygon

getgeometry

timberharvestedvolume

harvestablefuelwood
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/VincyaneBadouard/LoggingLab.git}
  \item Report bugs at \url{https://github.com/VincyaneBadouard/LoggingLab/issues}
}

}
\author{
\strong{Maintainer}: Vincyane Badouard \email{vincyane.badouard@gmail.com}

Authors:
\itemize{
  \item Sylvain Schmitt \email{sylvain.m.schmitt@gmail.com}
  \item Guillaume Salzet \email{Guillaume.Salzet@ecofog.gf}
  \item Thomas Gaquiere \email{gaquiere.thomas@yahoo.com}
  \item Géraldine Derroire \email{geraldine.derroire@cirad.fr}
}

}
\keyword{internal}
