% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingparameters.R
\name{loggingparameters}
\alias{loggingparameters}
\title{Other parameters of the logging simulator}
\usage{
loggingparameters(
  MinDBHValue = 10,
  MaxAreaSlope = 27,
  MaxTrailCenterlineSlope = 22,
  MaxTrailCrossSlope = 4,
  GrappleMaxslope = 20,
  TreeMaxSlope = 22,
  PlateauMaxSlope = 5,
  SlopeDistance = 6,
  WaterSourcesBufferZone = 30,
  MinMainTrailWidth = 5,
  MaxMainTrailWidth = 6,
  ScndTrailWidth = 4,
  BigTrees = 50,
  ObjectiveBonus = 30,
  CableLength = 40,
  GrappleLength = 6,
  IsolateTreeMinDistance = 100,
  FutureTreesMinDiameter = 35,
  TreefallSuccessProportion = 0.6,
  MinTreefallOrientation = 30,
  MaxTreefallOrientation = 45,
  TreeHollowPartForFuel = 1/3,
  Purge = 0.14,
  MaxTrailDensity = 200,
  MaxLandingArea = 1500,
  TreeHarvestableVolumeAllometry = function(DBH, aCoef, bCoef) aCoef + bCoef *
    (DBH/100)^2,
  TrunkHeightAllometry = function(DBH, TreeHarvestableVolume) TreeHarvestableVolume/(pi
    * (((DBH/100)/2)^2)),
  TreeHeightAllometry = function(DBH) exp(0.07359191 + 1.34241216 * log(DBH) +
    -0.12282344 * log(DBH)^2),
  CrownDiameterAllometry = function(DBH, TreeHeight, alpha, beta) exp(((log(DBH) -
    alpha - rnorm(length(DBH), 0, 0.0295966977))/beta))/TreeHeight,
  RottenModel = function(DBH) 1/(1 + exp(-(-5.151 + 0.042 * DBH * 100)))
)
}
\arguments{
\item{MinDBHValue}{Mimimum DBH value. Default = 10, in cm (double)}

\item{MaxAreaSlope}{Maximum area slope. Default = 27, in \% (double)}

\item{MaxTrailCenterlineSlope}{Maximum main&2nd trail centerline slope. Default = 22, in \% (double)}

\item{MaxTrailCrossSlope}{Maximum main&2nd trail cross slope. Default = 4, in \% (double)}

\item{GrappleMaxslope}{Grapple maximum slope. Default = 20, in \% (double)}

\item{TreeMaxSlope}{Trees slope. Default = 22, in \% (double)}

\item{PlateauMaxSlope}{Plateau maximum slope. Default = 5, in \% (double)}

\item{SlopeDistance}{Distance to compute slope. Default = 6, in m (3m for each side) (double)}

\item{WaterSourcesBufferZone}{Water sources buffer zone. Default = 30, in m (double)}

\item{MinMainTrailWidth}{Minimum main trail width. Default = 5, in m (double)}

\item{MaxMainTrailWidth}{Maximum main trail width. Default = 6, in m (double)}

\item{ScndTrailWidth}{2nd trail width. Default = 4, in m (double)}

\item{BigTrees}{Big trees. Default = 50, in cm (double)}

\item{ObjectiveBonus}{Objective bonus. Default = 30, in \% (20;30\%) (double)}

\item{CableLength}{Cable. Default = 40, in m (double)}

\item{GrappleLength}{Grapple. Default = 6, in m (double)}

\item{IsolateTreeMinDistance}{Isolate tree minimum distance. Default = 100, in m (double)}

\item{FutureTreesMinDiameter}{Future trees minimum diameter. Default = 35, in cm (double)}

\item{TreefallSuccessProportion}{TreefallSuccess proportion. Default = 0.6, (double)}

\item{MinTreefallOrientation}{Minimum treefall orientation. Default = 30, in degree (double)}

\item{MaxTreefallOrientation}{Maximum treefall orientation. Default = 45, in degree (double)}

\item{TreeHollowPartForFuel}{Part taken from hollow trees for fuel exploitation. Default = 1/3 (double)}

\item{Purge}{Purge. Default = 0.14, in m^3 of fuelwood/m^3 of logged trees (double)}

\item{MaxTrailDensity}{Maximum TrailDensity. Default = 200, in m/ha (double)}

\item{MaxLandingArea}{Maximum landing area. Default = 1500) in square meters (double)}

\item{TreeHarvestableVolumeAllometry}{(function)}

\item{TrunkHeightAllometry}{(function)}

\item{TreeHeightAllometry}{log(H) = 0.07359191 + 1.34241216 log(DBH) + -0.12282344 log(DBH)^2 (function)}

\item{CrownDiameterAllometry}{ln(D) = 𝜶+ 𝜷 ln(H*CD) + 𝜺, with 𝜺~N(0,σ^2) and meanσ^2 = 0.0295966977 (ref)(function)}

\item{RottenModel}{1 / (1 + exp(-(-5.151 + 0.042 DBH.100))) (function)}
}
\value{
A named list of 30 objects.
}
\description{
Other parameters of the logging simulator
}
\examples{
loggingparameters(MinDBHValue = 5)
}
