% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingsimulation.R
\name{loggingsimulation}
\alias{loggingsimulation}
\title{Logging simulation}
\usage{
loggingsimulation(
  inventory,
  plotmask,
  topography,
  creekdistances,
  speciescriteria,
  volumeparameters,
  scenario,
  objective = NULL,
  fuel = NULL,
  diversification = NULL,
  winching = NULL,
  directionalfelling = NULL,
  specieslax = FALSE,
  objectivelax = FALSE,
  crowndiameterparameters = ParamCrownDiameterAllometry,
  advancedloggingparameters = loggingparameters(),
  iter = 1,
  cores = 1
)
}
\arguments{
\item{inventory}{Input forest inventory for 1 plot and 1 census year (see the
inputs formats and metadata in the vignette or in
\code{\link{Paracou6_2016}}) (data.frame)
The columns required are:
\itemize{
\item \emph{Forest} (to apply the corresponding volume formula)
\item \emph{Plot} (1 value)
\item \emph{CensusYear} (1 value)
\item \emph{PlotArea}
\item \emph{idTree}
\item \emph{Xutm} and \emph{Yutm}
\item \emph{CodeAlive}
\item \emph{Family}, \emph{Genus}, \emph{Species} and \emph{VernName}
\item \emph{Circ} or \emph{CircCorr}
}}

\item{plotmask}{Inventoried plot mask
(SpatialPolygonsDataFrame \strong{with a crs in UTM})}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR,
1m resolution)
(RasterLayer \strong{with a crs in UTM}) (See \code{\link{DTMParacou}})
We advise you to generate your raster with Qgis rather than with the
'raster' package on R.}

\item{creekdistances}{Relative distances (vertical (\emph{distvert}) and horizontal
(\emph{disthorz})) (1 m resolution) from nearest channel network (list of 2 large
RasterLayers \strong{with a crs in UTM}) (See \code{\link{CreekDistances}})
To generate creek distances: \code{\link{CreekDistances}} in 'Articles'.}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in the
same format of \code{\link{SpeciesCriteria}} (2 levels of commercial
species) (data.frame)}

\item{volumeparameters}{Volume parameters table (in the same format of
\code{\link{ForestZoneVolumeParametersTable}}) to compute the harvestable
volume of each tree, depend to its geographic zone if several locations
(data.frame)}

\item{scenario}{Logging scenario among: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
vignette for details)}

\item{objective}{Objective volume (m3/ha) (numeric)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"
If fuel wood exploitation (fuel = "1" or "2") the tree will be recovered
from the crown with a grapple if possible (respected grapple conditions).
If not, recovery at the foot with a cable at an angle to the trail.
Avoid future/reserve trees if chosen.}

\item{diversification}{Possibility to log other species in addition to the
main commercial species (species with a value of 2 for commercial in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{winching}{Tree recovery =
"0": no cable or grapple (trail to tree foot)
"1": only cable (default = 40m)
"2": grapple (default = 6m) + cable (grapple priority)
If grapple + cable (winching = "2") without fuel wood (fuel = "0")
recovery of the tree foot with grapple if possible (respected grapple
conditions) otherwise with cable with angle to the trail.
Avoidance of future/reserves if chosen.}

\item{directionalfelling}{Directional felling =
"0": only to direct the foot of the tree towards the trail
"1": to direct the foot of the tree towards the trail + to avoid damage to
future and reserve trees if possible
"2": to avoid damage to future and reserve trees if possible
+ orientation angle to the trail. Among the 2 possible angle positions,
the position that favours the return to the main trail should be chosen.
The angle to the trail is favoured to avoid future/reserve trees.
If the avoidance of future/reserve trees could not be performed,
a message is returned.}

\item{specieslax}{Allow diversification if stand is too poor to reach the
objective volume without diversification, = FALSE by
default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the
objective volume (if stand too poor), = FALSE by default (logical)}

\item{crowndiameterparameters}{Crown diameter allometry parameters table (in
the same format of \code{\link{ParamCrownDiameterAllometry}}) to compute the
crown diameter of each tree, depend to its DBH (Diameter at Breast Height)
and its Species, Genus or Family names (Aubry-Kientz et al.2019).
(data.frame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}

\item{iter}{Number of iterations (numeric). Default = 1.}

\item{cores}{Number of cores for parallelization (numeric). Default = 1.}
}
\value{
A large list of 38 elements for each iteration, contained in a list.
Input inventory (data.frame) with logging informations (list) (see the
outputs metadata in the vignette or
\code{\link{LoggingSimulationOutputs_iter}}).
}
\description{
This function allows to simulate a timber and fuel wood logging
on a forest plot. It covers: harvestable zones definition, tree selection,
secondary skidding trails layout, tree felling, timber harvested, fuel wood
volume and short-term damages quantification. This simulator is
individual-centred, spatialised, and takes into account the topography and
the hydrographic network.
}
\examples{
\dontrun{
data(Paracou6_2016) # inventory
data(DTMParacou) # topography
data(CreekDistances) # relative elevation
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters
data(ParamCrownDiameterAllometry) # parameters values of the crown diameter allometry

Rslt <- loggingsimulation(
 Paracou6_2016, plotmask = PlotMask, topography = DTMParacou,
 creekdistances  = CreekDistances, speciescriteria = SpeciesCriteria,
 volumeparameters = ForestZoneVolumeParametersTable, scenario = "manual",
 objective = 10, fuel = "2", diversification = TRUE, winching = "2",
 directionalfelling = "2", specieslax = FALSE, objectivelax = TRUE,
 crowndiameterparameters = ParamCrownDiameterAllometry,
 advancedloggingparameters = loggingparameters(), iter = 2, cores = 2)
 }

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{ForestZoneVolumeParametersTable}},
\code{\link{ParamCrownDiameterAllometry}}, \code{\link{loggingparameters}},

\code{\link{addtreedim}}, \code{\link{treeselection}},
\code{\link{harvestable}}, \code{\link{selected}},
\code{\link{futurereserve}}, \code{\link{treefelling}},
\code{\link{createcanopy}}, \code{\link{treefromthesky}},
\code{\link{felling1tree}}, \code{\link{rotatepolygon}},
\code{\link{getgeometry}}, \code{\link{timberharvestedvolume}},
\code{\link{exploitablefuelwoodvolume}}
}
