% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingsimulation1.R
\name{loggingsimulation1}
\alias{loggingsimulation1}
\title{Logging simulation without iteration or parallelization}
\usage{
loggingsimulation1(
  inventory,
  plotmask,
  topography,
  creekdistances,
  speciescriteria,
  volumeparameters,
  scenario,
  objective = NULL,
  fuel = NULL,
  diversification = NULL,
  winching = NULL,
  directionalfelling = NULL,
  specieslax = FALSE,
  objectivelax = FALSE,
  crowndiameterparameters = ParamCrownDiameterAllometry,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (Default: \code{\link{DTMParacou}}) (RasterLayer)}

\item{creekdistances}{Relative distances (vertical and horizontal) (1 m
resolution) from nearest channel network (list of 2 large RasterLayers)
(Default: \code{\link{CreekDistances}})
To generate vertical creek height:
\code{\link{CreekDistances}} in 'docs' folder of the package}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in the
same format of \code{\link{SpeciesCriteria}} (2 levels of commercial
species) (data.frame)}

\item{volumeparameters}{Volume parameters table (in the same format of
\code{\link{ForestZoneVolumeParametersTable}}) to compute the harvestable
volume of each tree, depend to its geographic zone if several locations
(data.frame)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{objective}{Objective volume (m3/ha) (numeric)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"
If fuel wood exploitation (fuel = "1" or "2") the tree will be recovered
from the crown with a grapple if possible (respected grapple conditions).
If not, recovery at the foot with a cable at an angle to the trail.
Avoid future/reserve trees if chosen.}

\item{diversification}{Possibility to log other species in addition to the
main commercial species (species with a value of 2 for commercial in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{winching}{"0": no cable or grapple (trail to tree foot)
"1": only cable (default = 40m)
"2": grapple (default = 6m) + cable (grapple priority)
If grapple + cable (winching = "2") without fuel wood (fuel = "0")
recovery of the tree foot with grapple if possible (respected grapple
conditions) otherwise with cable with angle to the trail.
Avoidance of future/reserves if chosen.}

\item{directionalfelling}{Directional felling =
"0": only to direct the foot of the tree towards the trail
"1": to direct the foot of the tree towards the trail + to avoid damage to
future and reserve trees if possible
"2": to avoid damage to future and reserve trees if possible
+ orientation angle to the trail. Among the 2 possible angle positions,
the position that favours the return to the main trail should be chosen.
The angle to the trail is favoured to avoid future/reserve trees.
If the avoidance of future/reserve trees could not be performed,
a message is returned.}

\item{specieslax}{Allow diversification if stand is too poor to reach the
objective volume without diversification, = FALSE by
default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the
objective volume (if stand too poor), = FALSE by default (logical)}

\item{crowndiameterparameters}{Crown diameter allometry parameters table (in
the same format of \code{\link{ParamCrownDiameterAllometry}}) to compute the
crown diameter of each tree, depend to its DBH (Diameter at Breast Height)
and its species, genus or family. (data.frame)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
Input inventory (data.frame) with logging informations (list)
(see the outputs metadata in the \code{\link{vignette}}).
}
\description{
This function allows to simulate a timber and wood energy logging
on a forest plot. It covers: harvestable area definition, tree selection,
main (only for ONF plots) and secondary trails layout, tree felling,
landings implementation (only for ONF plots), timber harvested, fuel wood
volume and damages quantification.
}
\examples{
data(Paracou6_2016) # inventory
data(PlotMask) # inventoried plot mask
data(DTMParacou) # topography
data(CreekDistances) # relative elevation
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters
data(ParamCrownDiameterAllometry) # parameters values of the crown diameter allometry

Rslt <- loggingsimulation1(
 Paracou6_2016, plotmask = PlotMask, topography = DTMParacou,
 creekdistances  = CreekDistances, speciescriteria = SpeciesCriteria,
 volumeparameters = ForestZoneVolumeParametersTable, scenario = "manual",
 objective = 10, fuel = "2", diversification = TRUE, winching = "2",
 directionalfelling = "2", specieslax = FALSE, objectivelax = TRUE,
 crowndiameterparameters = ParamCrownDiameterAllometry,
 advancedloggingparameters = loggingparameters())

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{ForestZoneVolumeParametersTable}},
\code{\link{ParamCrownDiameterAllometry}}, \code{\link{loggingparameters}},

\code{\link{inventorycheckformat}}, \code{\link{cleaninventory}}
\code{\link{addtreedim}}, \code{\link{treeselection}},
\code{\link{harvestable}}, \code{\link{selected}},
\code{\link{futurereserve}}, \code{\link{treefelling}},
\code{\link{createcanopy}}, \code{\link{treefromthesky}},
\code{\link{felling1tree}}, \code{\link{rotatepolygon}},
\code{\link{getgeometry}}, \code{\link{timberharvestedvolume}},
\code{\link{exploitablefuelwoodvolume}}
}
