% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingsummary.R
\name{loggingsummary}
\alias{loggingsummary}
\title{Summary of the outputs of the loggingsimulation function}
\usage{
loggingsummary(x)
}
\arguments{
\item{x}{Outputs of the loggingsimulation function (list)}
}
\value{
A summary of the outputs of the loggingsimulation function
(console display):
\itemize{
\item a reminder of the inputs
\item the iterations statistics (mean, standard deviation, quantiles) of the
outgoing figures
Please note that all volumes in m3/ha and biomass in ton/ha are per
exploitable hectare, not per plot hectare.
}
}
\description{
Summary of the outputs of the loggingsimulation function
}
\examples{
data(LoggingSimulationOutputs_iter) # Outputs of one logging simulation (2 iterations and 2 cores)

loggingsummary(x = LoggingSimulationOutputs_iter)

}
