% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingsummary1.R
\name{loggingsummary1}
\alias{loggingsummary1}
\title{Summary of the outputs of the loggingsimulation1 function}
\usage{
loggingsummary1(x)
}
\arguments{
\item{x}{Outputs of the loggingsimulation function (list)}
}
\value{
A summary of the outputs of the loggingsimulation1 function
(console display):
\itemize{
\item a reminder of the inputs
\item the outgoing figures
\item a print of the output inventory
Please note that all volumes in m3/ha and biomass in ton/ha are per
exploitable hectare, not per plot hectare.
}
}
\description{
Summary of the outputs of the loggingsimulation1 function
}
\examples{
data(LoggingSimulationOutputs) # Outputs of one logging simulation

loggingsummary1(LoggingSimulationOutputs)

}
