% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loggingsummary1.R
\name{loggingsummary1}
\alias{loggingsummary1}
\title{Summary of the outputs of the loggingsimulation1 function}
\usage{
loggingsummary1(x)
}
\arguments{
\item{x}{Outputs of the loggingsimulation function (list)}
}
\value{
A summary of the outputs of the loggingsimulation1 function
(console display):
\itemize{
\item a reminder of the inputs
\item the outgoing figures
\item a print of the output inventory
}
}
\description{
Summary of the outputs of the loggingsimulation1 function
}
\examples{
\dontrun{
data(Paracou6_2016) # inventory
data(PlotMask) # inventoried plot mask
data(DTMParacou) # topography
data(VerticalCreekHeight) # relative elevation
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters
data(ParamCrownDiameterAllometry) # parameters values of the crown diameter allometry

Rslt <- suppressMessages(loggingsimulation1(
 Paracou6_2016, plotmask = PlotMask, topography = DTMParacou,
 verticalcreekheight  = VerticalCreekHeight, speciescriteria = SpeciesCriteria,
 volumeparameters = ForestZoneVolumeParametersTable, scenario = "manual",
 objective = 20, fuel = "2", diversification = TRUE, winching = "2",
 directionalfelling = "2", specieslax = FALSE, objectivelax = TRUE,
 crowndiameterparameters = ParamCrownDiameterAllometry,
 advancedloggingparameters = loggingparameters())
 )

loggingsummary1(Rslt)
}

}
