% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MainTrails.R
\docType{data}
\name{MainTrails}
\alias{MainTrails}
\title{Main trails of Paracou plot 6}
\format{
sf (sfc_LINESTRING)
}
\usage{
MainTrails
}
\description{
The main trails are defined as lines at the edge and outside of the plot
(plot 6 of Paracou, French Guiana)
}
\keyword{datasets}
