% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Maria.R
\docType{package}
\name{Maria}
\alias{Maria}
\title{Maria-package}
\description{
Logging simulator
}
\details{
This package allows to simulate the different low impact logging versions in French Guiana.

The Maria package provides n functions:
treeselection,  and .
}
\section{Maria functions}{

loggingsimulation

loggingparameters
scenariosparameters

inventorycheckformat
addtreedim

treeselection
ONFGuyafortaxojoin
harvestable
selected
futurereserve

treefelling
createcanopy
treefromthesk
directionalfellingsuccessdef
felling1tree
rotatepolygon
getgeometry

timberharvestedvolume
exploitablefuelwoodvolume
}

