% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONFGuyafortaxojoin.R
\name{ONFGuyafortaxojoin}
\alias{ONFGuyafortaxojoin}
\title{ONFGuyafortaxojoin}
\usage{
ONFGuyafortaxojoin(inventory, speciescriteria)
}
\arguments{
\item{inventory}{your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in
the same format of \code{\link{SpeciesCriteria}} (data.frame)}
}
\value{
The forest inventory with economic informations (economic name, rank,
mimimum/maximum falling diameter, upped mimimum falling diameter) per
individuals (data.frame)
}
\description{
Joins a commercial criteria table to a forest inventory
}
\examples{
data(Paracou6_2016)
data(SpeciesCriteria)
ONFGuyafortaxojoin(Paracou6_2016, SpeciesCriteria)

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}}
}
