% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ONFGuyafortaxojoin.R
\name{ONFGuyafortaxojoin}
\alias{ONFGuyafortaxojoin}
\title{Joins commercial criteria table to the forest inventory}
\usage{
ONFGuyafortaxojoin(inventory, speciescriteria)
}
\arguments{
\item{inventory}{input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in
the same format as \code{\link{SpeciesCriteria}} (data.frame)}
}
\value{
The forest inventory with economic informations (commercial name, economic interest level,
minimum/maximum felling diameter, increased minimum felling diameter) added for each
individual in the variables ONFName, Commercial, MinFD, UpMinFD, MaxFD (data.frame)
}
\description{
Joins a commercial criteria table to a forest inventory
}
\examples{
data(Paracou6_2016)
data(SpeciesCriteria)
new <- ONFGuyafortaxojoin(Paracou6_2016, SpeciesCriteria)

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}}
}
