% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMask.R
\docType{data}
\name{PlotMask}
\alias{PlotMask}
\title{Paracou P6 plot mask}
\format{
SpatialPolygonsDataFrame \strong{with a crs in UTM}
}
\usage{
PlotMask
}
\description{
Mask of the plot 6 of Paracou (1m resolution, LiDAR campaign of 2016) (French Guiana)
}
\keyword{datasets}
