% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSlope.R
\docType{data}
\name{PlotSlope}
\alias{PlotSlope}
\title{PlotSlope}
\format{
SpatialPolygonsDataFrame
}
\usage{
PlotSlope
}
\description{
Slopes (in radians) of plot 6 of Paracou (French Guiana) (calculated with a neighbourhood of 8 cells)
}
\keyword{datasets}
