% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenariosparameters.R
\name{scenariosparameters}
\alias{scenariosparameters}
\title{Parameters of the scenarios}
\usage{
scenariosparameters(
  scenario,
  objective = NULL,
  fuel = NULL,
  diversification = NULL,
  winching = NULL,
  directionalfelling = NULL
)
}
\arguments{
\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{objective}{Objective volume per hectare (numeric)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"}

\item{diversification}{Possibility to log other species in addition to the main
commercial species (species with a value of 2  for commercial in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{winching}{No cable or grapple = "0", cable only = "1", grapple + cable =
"2"}

\item{directionalfelling}{Directional felling = "0" (absent), "1" (only to
avoid damage to future and reserve trees), "2" (to avoid damage to future and
reserve trees and to position the log relatively to the track)}
}
\value{
A named list of 5 objects.
}
\description{
Parameters of the scenarios
}
\examples{
scenariosparameters(scenario = "RIL1")

}
