% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SecondTrailsOpening.R
\name{secondtrailsopening}
\alias{secondtrailsopening}
\title{Second trails opening}
\usage{
secondtrailsopening(
  topography,
  plotmask,
  creekdistances,
  treeselectionoutputs,
  CostMatrix = list(list(list(Slope = 3, Cost = 3), list(Slope = 5, Cost = 5),
    list(Slope = 12, Cost = 20), list(Slope = 22, Cost = 60), list(Slope = 27, Cost =
    600), list(Slope = Inf, Cost = 1000)), list(list(CostType = "Initial", CostValue =
    1000), list(CostType = "Access", CostValue = 1000), list(CostType = "BigTrees",
    CostValue = 500), list(CostType = "Reserves", CostValue = 500), list(CostType =
    "Futures", CostValue = 50), list(CostType = "MainTrails", CostValue = 1e-04),
    list(CostType = "SecondTrails", CostValue = 0.1))),
  scenario,
  winching = NULL,
  fact = 3,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{creekdistances}{Relative distances (vertical (distvert) and horizontal
(distHorz)) from nearest channel network (list of 2 large RasterLayers)}

\item{treeselectionoutputs}{A list with:
\itemize{
\item your inventory with: "DistCriteria", "Slope", "SlopeCriteria", "LoggingStatus",
"Selected", "Up", "VolumeCumSum", "ProbedHollowProba", "ProbedHollow"
new columns (see the outputs metadata in the \code{\link{vignette}}).
\item the objective volume with or without a bonus (if hollow trees exploitation)
\item 6 sets of spatial points:
harvestable, selected, future and reserve, hollow and fuel wood trees
}}

\item{CostMatrix}{List of list defining conditional weight over binned slopes
values}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{winching}{"0": no cable or grapple (trail to tree foot)
"1": only cable (default = 40m)
"2": grapple (default = 6m) + cable (grapple priority)
If grapple + cable (winching = "2") without fuel wood (fuel = "0")
recovery of the tree foot with grapple if possible (respected grapple
conditions) otherwise with cable with angle to the trail.
Avoidance of future/reserves if chosen.}

\item{fact}{Aggregation factor of cost raster resolution to initial DTM}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with :
\itemize{
\item RawSecondTrails : non-smoothed secondary trails
\item TrailsIdentity: information on sections of the trails with:
\itemize{
\item LineID:
\item LoggedTrees: idTree of trees reached by the trails
\item TypeExpl: type of winching
}
\item SmoothedSecondTrails: Smoothed secondary trails polygons (sfc_MULTIPOLYGON)
\item TrailsDensity: Second trails density (in m/ha)
\item inventory: Updated inventory
\item CostRasterAgg: A cost Raster (RasterLayer)
}
}
\description{
Second trails opening
}
\examples{
\dontrun{
data(Paracou6_2016)
data(PlotMask)
data(DTMParacou)
data(HarvestablePolygons)
data(MainTrails)
data(PlotSlope)
data(SpeciesCriteria)
data(CreekDistances)

inventory <- commercialcriteriajoin(addtreedim(cleaninventory(Paracou6_2016, PlotMask),
 volumeparameters = ForestZoneVolumeParametersTable),SpeciesCriteria)


treeselectionoutputs <- treeselection(inventory,
topography = DTMParacou, plotslope = PlotSlope, MainTrails = MainTrails,
harvestablepolygons = HarvestablePolygons,
speciescriteria = SpeciesCriteria,
scenario = "RIL3",
objectivelax = TRUE,
advancedloggingparameters = loggingparameters())

secondtrails <- secondtrailsopening(
  topography = DTMParacou,
  plotmask = PlotMask,
  treeselectionoutputs = treeselectionoutputs,
  creekdistances = CreekDistances,
  CostMatrix = list(list(list(Slope = 3, Cost = 3),
                         list(Slope = 5, Cost = 5),
                         list(Slope = 12, Cost = 20),
                         list(Slope = 22, Cost = 100),
                         list(Slope = 27, Cost = 1000),
                         list(Slope = Inf, Cost = Inf)),
                    list(list(CostType = "Initial", CostValue = 1000),
                         list(CostType = "Access", CostValue = Inf),
                         list(CostType = "BigTrees", CostValue = 500),
                         list(CostType = "Reserves", CostValue = 1000),
                         list(CostType = "Futures", CostValue = 900),
                         list(CostType = "MainTrails", CostValue = 1E-4),
                         list(CostType = "SecondTrails", CostValue = 1E-2))),
  scenario = "RIL3",
  fact = 3,
  advancedloggingparameters = loggingparameters())

library(ggplot2)
library(sf)
ggplot() +
  # Harvestable zones
  geom_sf(data = HarvestablePolygons,
        aes(alpha = Harvestable),
        fill = "olivedrab") +
  labs(alpha = "Harvestable") +
  labs(title = "P6 zones exploitables") +

  # 2ndary trails
  geom_sf(data = st_as_sf(secondtrails[[3]]), col = "darkgreen") +
  geom_sf(data = st_as_sf(secondtrails[[1]]), col = "red")

secondtrails[[4]]
  }

}
