% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{selected}
\usage{
selected(
  inventory,
  type = "manual",
  fuel,
  diversification,
  specieslax = FALSE,
  objectivelax = FALSE,
  DEM = DemParacou,
  otherloggingparameters = loggingparameters(),
  VO,
  HVinit
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{type}{"RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel",
"RIL3fuelhollow" or "manual"(character)}

\item{fuel}{no  exploitation = "0" (default), damage exploitation in fuelwood
= "1", exploitation of hollow trees and damage in fuelwood = "2"}

\item{diversification}{(logical)}

\item{specieslax}{= FALSE by default (logical)}

\item{objectivelax}{= FALSE by default (logical)}

\item{DEM}{(RasterLayer)}

\item{otherloggingparameters}{(list)}

\item{VO}{(numeric value)}

\item{HVinit}{(numeric value)}
}
\value{
A FAIRE
}
\description{
selected
}
\examples{

data(Paracou6_2016)
data(DemParacou)

inventory <- ONFGuyafortaxojoin(addtreedim(cleaninventory(
inventorycheckformat(Paracou6_2016))))

harvestableOutputs <- harvestable(inventory, diversification = TRUE,
 specieslax = FALSE,
DEM = DemParacou, plotslope = PlotSlope,
otherloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

selecInventory <- selected(inventory, type = "manual", fuel = "2",
diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters(), VO = 30,
HVinit = HVinit)$inventory

}
