% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{selected}
\usage{
selected(
  inventory,
  scenario = "manual",
  fuel,
  diversification,
  specieslax = FALSE,
  objectivelax = FALSE,
  DEM,
  advancedloggingparameters = loggingparameters(),
  VO,
  HVinit
)
}
\arguments{
\item{inventory}{your inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", damage
exploitation in fuelwood = "1", exploitation of hollow trees and damage in
fuelwood = "2"}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by
default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the
objective volume (if stand too poor), = FALSE by default (logical)}

\item{DEM}{Digital terrain model (DTM) of the inventoried plot (LiDAR or
SRTM) (default: \code{\link{DemParacou}}) (RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}

\item{VO}{The objective volume with or without a bonus (if hollow trees
exploitation)(numeric value) (see the \code{\link{loggingparameters}})}

\item{HVinit}{the harvestable volume in the plot for your criteria
(\code{\link{SpeciesCriteria}}) (numeric value)}
}
\value{
Your inventory with the trees selected for harvesting (depending on
the logging scenario chosen), and 2 sets of spatial points: (HollowTrees
and EnergywoodTrees)
}
\description{
selected
}
\examples{

data(Paracou6_2016)
data(DemParacou)
data(PlotSlope)

inventory <- ONFGuyafortaxojoin(addtreedim(
inventorycheckformat(Paracou6_2016)))

harvestableOutputs <- harvestable(inventory, diversification = TRUE,
 specieslax = FALSE,
DEM = DemParacou, plotslope = PlotSlope,
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

selecInventory <- selected(inventory, scenario = "manual", fuel = "2",
diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
DEM = DemParacou, advancedloggingparameters = loggingparameters(), VO = 30,
HVinit = HVinit)$inventory

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{DemParacou}}, \code{\link{loggingparameters}}
}
