% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{Trees to be exploited selection}
\usage{
selected(
  inventory,
  topography,
  scenario,
  fuel = NULL,
  diversification = NULL,
  VO,
  HVinit,
  specieslax = FALSE,
  objectivelax = FALSE,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", damages and purge
exploitation in fuelwood = "1", exploitation of hollow trees, damages and
purge in fuelwood = "2"}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{VO}{The objective volume for the entire surface of the plot (numeric
value)}

\item{HVinit}{the harvestable volume in the plot for chosen criteria
(\code{\link{SpeciesCriteria}}) (numeric value)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by
default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the
objective volume (if stand too poor), = FALSE by default (logical)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
Input inventory with the trees selected for harvesting (depending on
the logging scenario chosen), and 2 sets of spatial points: (HollowTrees and
EnergywoodTrees)
}
\description{
Trees to be exploited selection
}
\examples{
data(Paracou6_2016)
data(DTMParacou)
data(PlotSlope)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- ONFGuyafortaxojoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
diversification = TRUE, specieslax = FALSE, plotslope = PlotSlope,
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

selecInventory <- selected(inventory, topography = DTMParacou,
scenario = "manual", fuel = "2", diversification = TRUE,
VO = 125, HVinit = HVinit, specieslax = FALSE, objectivelax = TRUE,
advancedloggingparameters = loggingparameters())$inventory

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{DTMParacou}}, \code{\link{loggingparameters}}
}
