% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{selected}
\usage{
selected(
  inventory,
  type,
  fuel = "0",
  diversification,
  specieslax = FALSE,
  objectivelax = FALSE,
  otherloggingparameters = loggingparameters(),
  VO,
  HVinit
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{type}{"RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel", "RIL3fuelhollow" or "manual"(character)}

\item{fuel}{no  exploitation = "0" (default),
damage exploitation in fuelwood = "1", exploitation of hollow trees and damage in fuelwood = "2"}

\item{diversification}{(logical)}

\item{specieslax}{= FALSE by default (logical)}

\item{objectivelax}{= FALSE by default (logical)}

\item{otherloggingparameters}{(list)}

\item{VO}{(numeric value)}

\item{HVinit}{(numeric value)}
}
\value{

}
\description{
selected
}
\examples{
inventory = harvestable(ONFGuyafortaxojoin(addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))),
diversification = TRUE, specieslax = FALSE)$inventory
HVinit = harvestable(ONFGuyafortaxojoin(addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))),
diversification = TRUE, specieslax = FALSE)$HVinit
selected(inventory, type = "manual", fuel = "0", diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters(), VO = 20, HVinit = HVinit)

}
