% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selected.R
\name{selected}
\alias{selected}
\title{Selection of trees to log}
\usage{
selected(
  inventory,
  topography,
  scenario,
  fuel = NULL,
  diversification = NULL,
  VO,
  HVinit,
  specieslax = FALSE,
  objectivelax = FALSE,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"}

\item{diversification}{TPossibility to log other species in addition to the
main commercial species (species with a value of 2  for commercial in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{VO}{Objective volume for the entire area of the plot (numeric value)}

\item{HVinit}{Harvestable volume in the plot for the chosen scenario
(\code{\link{SpeciesCriteria}}) (numeric value)}

\item{specieslax}{Allow diversification if the stand is too poor to reach the
objective volume without diversification, = FALSE by default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the
objective volume (if stand too poor), = FALSE by default (logical)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
Input inventory with
\itemize{
\item The trees selected for harvesting ("Selected")
\item if the Minimum Felling Diameter (MinFD) of their species has been raised
("Up")
\item The cumulative harvestable volume of harvestable trees("VolumeCumSum").
\item The probability of a tree being probed hollow ("ProbedHollowProba") and the
probed hollow trees ("ProbedHollow").
See the outputs metadata in the \code{\link{vignette}}
}

2 sets of spatial points: (HollowTrees and EnergywoodTrees)
}
\description{
Selection of trees to log
}
\details{
This function allows the selection of trees to be exploited among the
exploitable trees. It also informs about the trees that were finally found
to be hollow during the probe. If fuel = 2, the hollow trees will be
harvested and are therefore included in the objective volume. If fuel = 0 or
1, the hollow trees will not be exploited, so the function looks for other
trees to reach the objective volume (if possible).

If the harvestable volume is higher than the objective volume, MinFD of the
1st economic rank species is increased.
If this is not enough and if diversification is allowed, MinFD of 2nd
economic level species is increased. Then, the trees to be harvested are
chosen in decreasing order of volume, until the objective volume is reached.

If the harvestable volume is lower than the objective volume,
diversification can be applied if it was not already applied ('specieslax')
(trees of all commercial ranks are selected in decreasing order of volume
until the objective volume is reached), or harvesting can continue despite
an unreached objective volume, or be abandoned ('objectivelax')
}
\examples{
data(Paracou6_2016)
data(DTMParacou)
data(PlotSlope)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- ONFGuyafortaxojoin(inventory, SpeciesCriteria)

harvestableOutputs <- harvestable(inventory, topography = DTMParacou,
diversification = TRUE, specieslax = FALSE, plotslope = PlotSlope,
advancedloggingparameters = loggingparameters())

inventory <- harvestableOutputs$inventory
HVinit <- harvestableOutputs$HVinit

selecInventory <- selected(inventory, topography = DTMParacou,
scenario = "manual", fuel = "2", diversification = TRUE,
VO = 125, HVinit = HVinit, specieslax = FALSE, objectivelax = TRUE,
advancedloggingparameters = loggingparameters())$inventory

}
\seealso{
\code{\link{Paracou6_2016}}, \code{\link{SpeciesCriteria}},
\code{\link{DTMParacou}}, \code{\link{loggingparameters}}
}
