% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeCond.R
\name{SlopeRdCond}
\alias{SlopeRdCond}
\title{SlopeRdCond}
\usage{
SlopeRdCond(DEM, fact = 1, advancedloggingparameters = loggingparameters())
}
\arguments{
\item{DEM}{A RasterLayer for digital elevation model (DEM) data.}

\item{fact}{numeric (default = 3). The reduction factor of aggregated DEM.}

\item{advancedloggingparameters}{Other parameters of the logging simulator.}
}
\value{

}
\description{
SlopeRdCond
}
\examples{

data(DemParacou)

SlopeCond <- SlopeRdCond((DEM, fact=3))

}
