% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeRdCond.R
\name{SlopeRdCond}
\alias{SlopeRdCond}
\title{SlopeRdCond}
\usage{
SlopeRdCond(
  DTM,
  advancedloggingparameters = loggingparameters(),
  grapple = FALSE
)
}
\arguments{
\item{DTM}{A RasterLayer for digital terrain model (DTM) data}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}

\item{grapple}{Use grapple in harvest. Default = FALSE (boolean)}
}
\value{
A transition layer object with weighted adjacent graph according to
longitudinal / transversal slope condition
}
\description{
SlopeRdCond
}
\examples{
\dontrun{
data(DTMParacou)

SlopeCond <- SlopeRdCond(DTM = DTMParacou)
}

}
