% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloperdcond.R
\name{sloperdcond}
\alias{sloperdcond}
\title{sloperdcond}
\usage{
sloperdcond(
  topography,
  advancedloggingparameters = loggingparameters(),
  grapple = FALSE
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}

\item{grapple}{Use grapple in harvest. Default = FALSE (boolean)}
}
\value{
A transition layer object with weighted adjacent graph according to
longitudinal / transversal slope condition
}
\description{
sloperdcond
}
\examples{
\dontrun{
data(DTMParacou)

SlopeCond <- sloperdcond(topography = DTMParacou)
}

}
