% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SlopeRdCond.R
\name{sloperdcond}
\alias{sloperdcond}
\title{sloperdcond}
\usage{
sloperdcond(
  topography,
  directions = 4,
  advancedloggingparameters = loggingparameters(),
  grapple = FALSE
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer \strong{with a crs in UTM})
We advise you to generate your raster with Qgis rather than with the
'raster' package on R.}

\item{directions}{numeric (default = 4). The number of directions for
movement between cells, either 4 or 8.}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}

\item{grapple}{Use grapple in harvest. Default = FALSE (boolean)}
}
\value{
A transition layer object with weighted adjacent graph according to
longitudinal / transversal slope condition
}
\description{
sloperdcond
}
\examples{
\dontrun{
data(DTMParacou)

SlopeCond <- sloperdcond(topography = DTMParacou)
}

}
