% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothtrails.R
\name{smoothtrails}
\alias{smoothtrails}
\title{Trails smoothing}
\usage{
smoothtrails(
  paths,
  plotmask,
  verbose = FALSE,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{paths}{Raw secondary trails polygons (sp polylines)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{verbose}{return message on second trails density criteria (boolean)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with :
\enumerate{
\item SmoothedTrails: Smoothed trails polygons;
\item Smoothed second trails polygons;
\item Smoothed main trails polygons;
\item smoothed trail density.
}
}
\description{
Trails smoothing
}
\examples{
data(DTMParacou)
data(Paracou6_2016)
data(HarvestableAreaOutputsCable)
data(MainTrails)
data(PlotMask)
data(PlotSlope)
data("SpeciesCriteria")

scenarios <- scenariosparameters(scenario = "RIL3", objective = 15)

inventory <- commercialcriteriajoin(addtreedim(cleaninventory(Paracou6_2016, PlotMask),
 volumeparameters = ForestZoneVolumeParametersTable),SpeciesCriteria)

treeselectionoutputs <- treeselection(inventory,
topography = DTMParacou,
speciescriteria = SpeciesCriteria,
scenario = "manual", objective = 10, fuel = "2", diversification = TRUE,
winching = "0", specieslax = FALSE, objectivelax = TRUE,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
advancedloggingparameters = loggingparameters())

SecondTrailsRaw <- secondtrailsopening(DTM = DTMParacou, plotslope =  PlotSlope,plotmask = PlotMask,
  harvestablepolygons = HarvestablePolygons, maintrails = MainTrails, treeselectionoutputs,
  CostMatrix = list(list(list(Slope = 3, Cost = 3),
                           list(Slope = 5, Cost = 5),
                           list(Slope = 12, Cost = 20),
                           list(Slope = 22, Cost = 60),
                           list(Slope = 27, Cost = 600),
                           list(Slope = Inf, Cost = 1000)),
                      list(list(CostType = "Initial", CostValue = 1000),
                           list(CostType = "Access", CostValue = 1000),
                           list(CostType = "BigTrees", CostValue = 500),
                           list(CostType = "Reserves", CostValue = 500),
                           list(CostType = "Futures", CostValue = 50),
                           list(CostType = "MainTrails", CostValue = 1E-4),
                           list(CostType = "SecondTrails", CostValue = 0.1))),
 scenarios = scenarios,
 fact = 3,
 advancedloggingparameters = loggingparameters())

 SecondTrailsSmth <- smoothtrails(paths = SecondTrailsRaw[[1]],
                                 plotmask = PlotMask,
                                   )

}
