% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothtrails.R
\name{smoothtrails}
\alias{smoothtrails}
\title{Trails smoothing}
\usage{
smoothtrails(
  paths,
  pts,
  plots,
  partMainTrails,
  smthfact = 5,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{paths}{Raw secondary trails polygons (sp polylines)}

\item{pts}{Harvested tree locations (sf points)}

\item{plots}{Studied plots (sf)}

\item{partMainTrails}{Intersections between accessible area and MainTrails (sf
polygon)}

\item{smthfact}{A positive number controlling the smoothness and level of
generalization (numeric)}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with : 1: Smoothed secondary trails polygons.; 2: Second trails density
}
\description{
Trails smoothing
}
\examples{
\dontrun{
data(DTMParacou)
data(Paracou6_2016)
data(HarvestablePolygons)
data(MainTrails)
data(Plots)
data(PlotSlope)
data("SpeciesCriteria")

scenarios <- scenariosparameters(scenario = "RIL3", objective = 15)

inventory <- ONFGuyafortaxojoin(addtreedim(inventorycheckformat(Paracou6_2016),
 volumeparameters = ForestZoneVolumeParametersTable),SpeciesCriteria)

AccessPolygons <- FilterAccesExplArea(harvestablepolygons = HarvestablePolygons,
MainTrails = MainTrails,
winching = scenarios$winching,
advancedloggingparameters = loggingparameters())

treeselectionoutputs <- treeselection(inventory,
topography = DTMParacou,
speciescriteria = SpeciesCriteria, objective = scenarios$objective,
scenario ="manual", fuel = "2", diversification = TRUE, specieslax = FALSE,
objectivelax = TRUE, MainTrails = MainTrails, plotslope = PlotSlope,
harvestablepolygons = AccessPolygons,
advancedloggingparameters = loggingparameters())

SecondTrailsRaw <- secondtrailsopening(DTM = DTMParacou, plotslope =  PlotSlope,plots = Plots,
  harvestablepolygons = HarvestablePolygons, MainTrails = MainTrails, treeselectionoutputs,
  CostMatrix = list(list(list(Slope = 3, Cost = 3),
                           list(Slope = 5, Cost = 5),
                           list(Slope = 12, Cost = 20),
                           list(Slope = 22, Cost = 60),
                           list(Slope = 27, Cost = 600),
                           list(Slope = Inf, Cost = 1000)),
                      list(list(CostType = "Initial", CostValue = 1000),
                           list(CostType = "Access", CostValue = 1000),
                           list(CostType = "BigTrees", CostValue = 500),
                           list(CostType = "Reserves", CostValue = 500),
                           list(CostType = "Futures", CostValue = 50),
                           list(CostType = "MainTrails", CostValue = 1E-4),
                           list(CostType = "SecondTrails", CostValue = 0.1))),
 scenarios = scenarios,
 fact = 3,
 advancedloggingparameters = loggingparameters())

 SecondTrailsSmth <- smoothtrails(paths = SecondTrailsRaw[[1]],
                                 plots = Plots,
                                   )
}

}
