% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/st_temporaryintersection.R
\name{st_temporaryintersection}
\alias{st_temporaryintersection}
\title{st_temporaryintersection}
\usage{
st_temporaryintersection(
  topography,
  plotmask,
  treebuffers,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{topography}{Digital terrain model (DTM) of the inventoried plot (LiDAR
or SRTM) (\code{\link{DTMParacou}}) (RasterLayer)}

\item{plotmask}{Inventoried plot mask (SpatialPolygonsDataFrame)}

\item{treebuffers}{Polygons vector}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
sf_Polygons
}
\description{
st_temporaryintersection
}
\examples{
data(Paracou6_2016)
data(DTMParacou)
data(PlotMask)
data(SpeciesCriteria)
data(ForestZoneVolumeParametersTable)
data(HarvestableAreaOutputsCable)

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

treeselectionoutputs <- suppressWarnings(treeselection(
  inventory,
  topography = DTMParacou,
  speciescriteria = SpeciesCriteria,
  scenario = "manual", objective = 20,
  fuel = "0",
  diversification = FALSE,
  winching = "0",
  specieslax = FALSE, objectivelax = TRUE,
  plotslope = HarvestableAreaOutputsCable$PlotSlope,
  harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
  advancedloggingparameters = loggingparameters()))

treebuffers <- treeselectionoutputs$SelectedTreesPoints \%>\% sf::st_buffer(40)

st_temporaryintersection(DTMParacou, PlotMask, treebuffers)

}
