% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timberharvestedvolume.R
\name{timberharvestedvolume}
\alias{timberharvestedvolume}
\title{Compute the timber harvested volume}
\usage{
timberharvestedvolume(
  inventory,
  scenario,
  fuel = NULL,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
vignette) (data.frame)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
vignette)}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with the logged volume (\emph{TimberLoggedVolume})
and when fuel = "2", the logged volume without the hollow trees
(\emph{NoHollowTimberLoggedVolume}).
\emph{NoHollowTimberLoggedVolume} is the logged volume (\emph{TimberLoggedVolume})
when fuel = "0" or "1".
}
\description{
Computes the timber harvested volume in healthy trees exploited
for timber, and in the hollow trees if they are also harvested for fuel
wood.
}
\details{
When fuel is "2", by default, 2/3 of the log of a hollow tree will
be usable as timber and 1/3 as fuel wood ('\emph{TreeHollowPartForFuel}'). When
fuel is not "2", hollow trees are not used for timber nor fuel wood. Dead
trees from the operation (trails, secondary windfall) are never exploited
as timber, but as fuel wood if chosen.
}
\examples{
data(Paracou6_2016) # inventory
data(DTMParacou) # topography
data(MainTrails)  # MainTrails
data(HarvestableAreaOutputsCable)
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters

inventory <- addtreedim(cleaninventory(Paracou6_2016, PlotMask),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- suppressMessages(treeselection(inventory,
topography = DTMParacou,
speciescriteria = SpeciesCriteria,
scenario = "manual", objective = 10, fuel = "2", diversification = TRUE,
winching = "0", specieslax = FALSE, objectivelax = TRUE,
harvestablearea = HarvestableAreaOutputsCable$HarvestableArea,
plotslope = HarvestableAreaOutputsCable$PlotSlope,
maintrails = MainTrails,
harvestablepolygons = HarvestableAreaOutputsCable$HarvestablePolygons,
advancedloggingparameters = loggingparameters())$inventory)

timberharvestedvolume(inventory, scenario = "manual", fuel = "2",
advancedloggingparameters = loggingparameters())

}
