% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timberharvestedvolume.R
\name{timberharvestedvolume}
\alias{timberharvestedvolume}
\title{Compute the timber harvested volume}
\usage{
timberharvestedvolume(
  inventory,
  scenario,
  fuel = NULL,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{Input inventory (see the inputs formats and metadata in the
\code{\link{vignette}}) (data.frame)}

\item{scenario}{Logging scenario: "RIL1", "RIL2broken", "RIL2", "RIL3",
"RIL3fuel", "RIL3fuelhollow" or "manual"(character) (see the
\code{\link{vignette}})}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", exploitation of
damage and unused part of logged trees for fuel = "1", exploitation of
hollow trees, damage and and unused part of the log for fuel = "2"}

\item{advancedloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list)}
}
\value{
A list with the logged volume (TimberLoggedVolume)
and when fuel = "2", the logged volume without the hollow trees
(NoHollowTimberLoggedVolume).
NoHollowTimberLoggedVolume is the logged volume (TimberLoggedVolume)
when fuel = "0" or "1".
}
\description{
Compute the timber harvested volume
}
\details{
When fuel is "2", by default, 2/3 of the log of a hollow tree will
be usable as timber and 1/3 as fuel wood ('TreeHollowPartForFuel'). When
fuel is not "2", hollow trees are not used for timber nor fuel wood. Dead
trees from the operation (trails, secondary windfall) are never exploited
as timber, but as fuel wood if chosen.
}
\examples{
data(Paracou6_2016) # inventory
data(DTMParacou) # topography
data(MainTrails)  # MainTrails
data(HarvestablePolygons) # HarvestablePolygons
data(SpeciesCriteria) # species exploitability criteria
data(ForestZoneVolumeParametersTable) # volume parameters

inventory <- addtreedim(inventorycheckformat(Paracou6_2016),
volumeparameters = ForestZoneVolumeParametersTable)

inventory <- suppressMessages(treeselection(inventory, objective = 20, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 harvestablepolygons = HarvestablePolygons, MainTrails = MainTrails,
 objectivelax = TRUE, topography = DTMParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory)

timberharvestedvolume(inventory, scenario = "manual", fuel = "2",
advancedloggingparameters = loggingparameters())

}
