% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timberharvestedvolume.R
\name{timberharvestedvolume}
\alias{timberharvestedvolume}
\title{timberharvestedvolume}
\usage{
timberharvestedvolume(
  inventory,
  fuel,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{fuel}{no  exploitation = "0", damage exploitation in fuelwood = "1",
exploitation of hollow trees and damage in fuelwood = "2"}

\item{otherloggingparameters}{(list)}
}
\value{
A list with the logged volume (LoggedVolume)
and when fuel = "2", the logged volume without the hollow trees (NoHollowLoggedVolume),
NoHollowLoggedVolume = NULL when fuel = "0" or "1
}
\description{
timberharvestedvolume
}
\examples{

\dontrun{
data(Paracou6_2016)

inventory <- addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))

inventory <- treeselection(inventory, SpeciesCriteria,
type ="manual", fuel = "2", objective = 20, diversification = TRUE, specieslax = FALSE,
objectivelax = FALSE, otherloggingparameters = loggingparameters())$inventory

inventory <- secondtrailsopening()$inventory

inventory <- treefelling()$inventory

inventory <- adjustedsecondtrails()$inventory

timberharvestedvolume(inventory, fuel = "2", otherloggingparameters = loggingparameters())
}

}
