% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefelling.R
\name{treefelling}
\alias{treefelling}
\title{treefelling}
\usage{
treefelling(
  inventory,
  type = "manual",
  fuel,
  directionalfelling,
  otherloggingparameters = loggingparameters(),
  AccessibleTreesPoints,
  SelectedTreesPoints,
  AvenirTreesPoints,
  ReserveTreesPoints,
  DeadTreesPoints,
  HollowTreesPoints,
  FuelwoodTreesPoints
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{type}{"RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel",
"RIL3fuelhollow" or "manual"(character)}

\item{fuel}{no  exploitation = "0", damage exploitation in fuelwood = "1",
exploitation of hollow trees and damage in fuelwood = "2"}

\item{directionalfelling}{directional felling = "0" (absent), "1" (only to
avoid damage to future and reserve trees), "2" (avoid damage to future and
reserve trees + track orientation)}

\item{otherloggingparameters}{(list)}

\item{AccessibleTreesPoints}{(MULTIPOINT)}

\item{SelectedTreesPoints}{(MULTIPOINT)}

\item{AvenirTreesPoints}{(MULTIPOINT)}

\item{ReserveTreesPoints}{(MULTIPOINT)}

\item{DeadTreesPoints}{(MULTIPOINT)}

\item{HollowTreesPoints}{(MULTIPOINT)}

\item{FuelwoodTreesPoints}{(MULTIPOINT) MainTrail 2ndTrail}
}
\value{
"Shadow" polygons +  "TreefallSuccess", "TreefallFailure",
"DamageTreesPoints", "DeadTreesPoints" vectors
}
\description{
treefelling
}
\examples{

\dontrun{
inventory <- addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))

inventory <- treeselection(inventory, SpeciesCriteria, type = "manual",
fuel = "0",objective = 20,
diversification = TRUE, specieslax = FALSE, objectivelax = FALSE,
otherloggingparameters = loggingparameters())$inventory

treefelling(inventory, type = "manual", fuel = "2", directionalfelling = "2",
 otherloggingparameters = loggingparameters())
 }

}
