% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treefelling.R
\name{treefelling}
\alias{treefelling}
\title{treefelling}
\usage{
treefelling(
  inventory,
  scenario = "manual",
  fuel,
  directionalfelling,
  MainTrail,
  ScndTrail,
  advancedloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{scenario}{"RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel",
"RIL3fuelhollow" or "manual"(character)}

\item{fuel}{no  exploitation = "0", damage exploitation in fuelwood = "1",
exploitation of hollow trees and damage in fuelwood = "2"}

\item{directionalfelling}{directional felling = "0" (absent), "1" (only to
avoid damage to future and reserve trees), "2" (avoid damage to future and
reserve trees + track orientation)}

\item{MainTrail}{(sfg)}

\item{ScndTrail}{(sfg)}

\item{advancedloggingparameters}{(list)}
}
\value{
"Shadow" polygons + "TreefallSuccess", "TreefallFailure",
"DamageTreesPoints", "DeadTreesPoints" vectors
}
\description{
treefelling
}
\examples{

\dontrun{

MainTrail <- sf::st_linestring(matrix(c(286400, 583130,
                                        286400, 583250,
                                        286655, 583250,
                                        286655, 583130,
                                        286400, 583130) # the return
                                     ,ncol=2, byrow=TRUE))

pol1 <- list(matrix(c(286503, 583134,
                      286503, 583240,
                      286507, 583240,
                      286507, 583134,
                      286503, 583134) # the return
                   ,ncol=2, byrow=TRUE))
pol2 <- list(matrix(c(286650, 583134,
                      286650, 583240,
                      286654, 583240,
                      286654, 583134,
                      286650, 583134) # the return
                   ,ncol=2, byrow=TRUE))

PolList = list(pol1,pol2) #list of lists of numeric matrices
ScndTrail <- sf::st_multipolygon(PolList)

inventory <- addtreedim(inventorycheckformat(Paracou6_2016))
inventory <- suppressMessages(treeselection(inventory, objective = 30, scenario ="manual",
 fuel = "2", diversification = TRUE, specieslax = FALSE,
 objectivelax = FALSE, topography = DTMParacou, plotslope = PlotSlope,
 speciescriteria = SpeciesCriteria,
 advancedloggingparameters = loggingparameters())$inventory)

inventory <- treefelling(inventory, scenario = "manual", fuel = "0",
directionalfelling = "2", MainTrail = MainTrail, ScndTrail = ScndTrail,
advancedloggingparameters = loggingparameters())

library(ggplot2)
ggplot() +
  geom_sf(data = st_as_sf(inventory, coords = c("Xutm", "Yutm"))) +
  geom_sf(data = getgeometry (inventory, TreePolygon), fill = "red") # trees polygons

st_intersection( # trees under the fallen trees
  getgeometry (inventory, TreePolygon),
  st_as_sf(inventory, coords = c("Xutm", "Yutm"))
) \%>\%
  ggplot() +
  geom_sf()
 }

}
