% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeselection.R
\name{treeselection}
\alias{treeselection}
\title{treeselection}
\usage{
treeselection(
  inventory,
  objective,
  type = "manual",
  fuel,
  diversification,
  specieslax = FALSE,
  objectivelax = FALSE,
  DEM = DemParacou,
  plotslope = PlotSlope,
  speciescriteria = SpeciesCriteria,
  otherloggingparameters = loggingparameters()
)
}
\arguments{
\item{inventory}{(data.frame)}

\item{objective}{Objective volume per hectare (numeric)}

\item{type}{Type of logging: "RIL1", "RIL2broken", "RIL2", "RIL3", "RIL3fuel",
"RIL3fuelhollow" or "manual"(character) (see the \code{\link{vignette}})}

\item{fuel}{Fuel wood exploitation: no exploitation = "0", damage exploitation
in fuelwood = "1", exploitation of hollow trees and damage in fuelwood = "2"}

\item{diversification}{Taking of other species in addition to the main
commercial species (2 levels of commercial species in the
\code{\link{SpeciesCriteria}} table) (logical)}

\item{specieslax}{Allow diversification if stand is too poor, = FALSE by
default (logical)}

\item{objectivelax}{Allow exploitation in case of non-achievement of the
objective volume (if stand too poor), = FALSE by default (logical)}

\item{DEM}{Digital terrain model (DTM) of the inventoried plot (LiDAR or
SRTM)(RasterLayer)}

\item{plotslope}{Slopes (in radians) of the inventoried plot (with a
neighbourhood of 8 cells) (RasterLayer)}

\item{speciescriteria}{Table of species exploitability criteria : species
names, economic interest level, minimum and maximum felling diameter, in the
same format of \code{\link{SpeciesCriteria}} (data.frame)}

\item{otherloggingparameters}{Other parameters of the logging simulator
\code{\link{loggingparameters}} (list) MainTrail (multiline)}
}
\value{
The new inventory with:
the commercial name (ONFName), the commercial level (Commercial), the "MinFD", "UpMinFD", "MaxFD",
"DistCrit", "Slope", "SlopeCrit", "LoggingStatus", "Selected", "Up", "VolumeCumSum", "ProbedHollowProba", "ProbedHollow"

the objective volume with or without a bonus (if
hollow trees exploitation), and the 6 set of specialised points:
harvestable, selected, future and reserve, hollow and energy wood trees
}
\description{
treeselection
}
\examples{

data(Paracou6_2016)
data(DemParacou)

inventory <- addtreedim(cleaninventory(inventorycheckformat(Paracou6_2016)))

treeselectionoutputs <- treeselection(inventory, objective = 20,
type ="manual", fuel = "2", diversification = FALSE, specieslax = FALSE,
objectivelax = FALSE, DEM = DemParacou, plotslope = PlotSlope,
speciescriteria = SpeciesCriteria,
otherloggingparameters = loggingparameters()) #  MainTrail


}
\seealso{
\code{\link{SpeciesCriteria}}
}
