% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VerticalCreekHeight.R
\docType{data}
\name{VerticalCreekHeight}
\alias{VerticalCreekHeight}
\title{Vertical creek height of Paracou plot 6}
\format{
A Large RasterLayer

To generate vertical creek height of any plot:
\code{\link{VerticalCreekHeight}} in 'docs' folder of the package
}
\usage{
VerticalCreekHeight
}
\description{
Vertical creek height (in m) of plot 6 of Paracou (French Guiana) (1 m resolution,
2016)
}
\keyword{datasets}
